% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{arima.string1}
\alias{arima.string1}
\title{Identify ARIMA model - internal function}
\usage{
arima.string1(object, padding = FALSE)
}
\arguments{
\item{object}{An object generate by Arima function}

\item{padding}{Logical.}
}
\description{
Identify ARIMA model - internal function
}
\keyword{internal}
