% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBackTesting_Plot.R
\name{change_model_factor_levels}
\alias{change_model_factor_levels}
\title{Change factor levels in order to get nice legends in ggplots}
\usage{
change_model_factor_levels(vect)
}
\arguments{
\item{vect}{A vector of the class \code{factor}.}
}
\description{
Change factor levels in order to get nice legends in ggplots
}
\examples{
x <- factor(c("MRWD", "LeeCarter", "HyndmanUllah", "Oeppen", "MEM6"))
new.x <- change_model_factor_levels(x)
}
\keyword{internal}
