% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_fit.R
\name{coda.input.check}
\alias{coda.input.check}
\title{Validate input values}
\usage{
coda.input.check(X)
}
\arguments{
\item{X}{A list with input arguments provided in \code{\link{coda}} function}
}
\description{
Validate input values
}
\keyword{internal}
