% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_fit.R
\name{coda}
\alias{coda}
\title{Fit CoDa Mortality Model}
\usage{
coda(data, x = NULL, y = NULL)
}
\arguments{
\item{data}{Matrix containing mortality data (dx) with ages as row and time as column.}

\item{x}{Vector of input ages (optional). Used to label the output objects and plots.}

\item{y}{Vector of input years (optional). Used to label the output objects and plots.}
}
\value{
The output is an object of class \code{"coda"} with the components:

\item{input}{List with arguments provided in input. Saved for convenience.}

\item{call}{An unevaluated function call, that is, an unevaluated 
expression which consists of the named function applied to the given arguments.}

\item{coefficients}{Estimated coefficients.}

\item{fitted.values}{Fitted values of the estimated CoDa model.}

\item{residuals}{Deviance residuals.}

\item{x}{Vector of ages used in the fitting.}

\item{y}{Vector of years used in the fitting.}
}
\description{
Fit Compositional Data (CoDa) model for forecasting the life table 
distribution of deaths. CoDa is a Lee-Carter type model. A key difference 
between the \insertCite{lee1992;textual}{MortalityForecast}
method and the Compositional Data model (CoDa) is that the former fits and 
forecasts the death rates (mx) while the latter is based on the life table 
death distribution (dx). 
\insertCite{@See @oeppen2008 and @bergeron2017;textual}{MortalityForecast} 
for a detail description and mathematical formulation.
}
\examples{
# Fit CoDa model
D <- MortalityForecast.data$dx
M <- coda(D)
summary(M)

# Forecast life expectancy
P <- predict(M, h = 20)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{predict.coda}}
}
