% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODA_predict.R
\name{compute_dx}
\alias{compute_dx}
\title{Internal function}
\usage{
compute_dx(dx, kt, ax, bx, fit, y, jumpchoice)
}
\arguments{
\item{kt}{Estimated kt vector of parameters}

\item{ax}{Estimated ax vector of parameters}

\item{bx}{Estimated bx vector of parameters}

\item{fit}{Fitted values}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}
}
\description{
Internal function
}
\keyword{internal}
