% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalRobustness.R
\name{computeRobustness}
\alias{computeRobustness}
\title{Compute robustness measures}
\usage{
computeRobustness(F1, F2, B1, B2, ND, measures, na.rm = TRUE)
}
\arguments{
\item{F1}{Forecast scenario 1.}

\item{F2}{Forecast scenario 2.}

\item{B1}{Forecast scenario 1 - benchmark.}

\item{B2}{Forecast scenario 2 - benchmark.}

\item{measures}{What robustness measure to compute? Various alternatives are 
available, \itemize{
 \item{Mean delta measures: } \code{"MD", "MAD", "sMRAD"};
 \item{Median delta measures: } \code{"MdD", "MdAD", "sMdRAD"}.
 }}

\item{na.rm}{A logical value indicating whether NA values should be stripped 
before the computation proceeds. Default: \code{TRUE}.}
}
\description{
Compute robustness measures
}
\author{
Marius D. Pascariu
}
\keyword{internal}
