% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMoments.R
\name{convertMoments}
\alias{convertMoments}
\title{Convert Statistical Moments}
\usage{
convertMoments(data, from = c("raw", "central", "normalized"),
  to = c("raw", "central", "normalized"), eta = NULL)
}
\arguments{
\item{data}{data.frame containing statistical moments.}

\item{from}{What type of statistical moments do we have in input \code{data}?
Three types of moments are accepted: \code{"raw", "central", "normalized"}.}

\item{to}{What type of statistical moments do you want to obtain? 
Three types of moments can be obtained: \code{"raw", "central", "normalized"}.}

\item{eta}{A numeric vector of the expected values. This is required ONLY 
is we convert central moments into raw or normalized. Default: \code{NULL}.}
}
\description{
Transform the raw, central or normalized statistical moments between them.
}
\details{
Wikipedia: In probability theory and statistics, 
the standardized moment of a probability distribution is a moment 
(normally a higher degree central moment) that is normalized. 
The normalization is typically a division by an expression of the 
standard deviation which renders the moment scale invariant. 
This has the advantage that such normalized moments differ only in other 
properties than variability, facilitating e.g. comparison of shape of 
different probability distributions.
}
\examples{
# raw moments
RM <- c(1, 68.75099, 4991.724, 371531.9, 28199680,  
        2176435499, 170477697491)


CM1 <- convertMoments(RM, from = "raw", to = "central")    # raw to central
NM1 <- convertMoments(RM, from = "raw", to = "normalized") # raw to normalized

CM2 <- convertMoments(NM1, from = "normalized", to = "central")
RM2 <- convertMoments(NM1, from = "normalized", to = "raw")

RM3 <- convertMoments(CM2, from = "central", to = "raw", eta = 68.75099)
NM3 <- convertMoments(CM2, from = "central", to = "normalized", eta = 68.75099)

# The resulted error following multiple conversions is negligible
sum(RM - RM3)
}
