% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMoments.R
\name{convertNM}
\alias{convertNM}
\title{Convert Normalized Moments to Central or Raw moments}
\usage{
convertNM(data, Out = c("central", "raw"))
}
\arguments{
\item{data}{Data.frame with normalized moments.}

\item{Out}{What type of statistical moments do we want to get? 
Three types moments can be obtained: \code{"raw", "central", "normalized"}.}
}
\description{
Convert Normalized Moments to Central or Raw moments
}
\keyword{internal}
