% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMoments.R
\name{convertRM}
\alias{convertRM}
\title{Convert Raw Moments to Central or Normalized moments}
\usage{
convertRM(data, Out = c("central", "normalized"))
}
\arguments{
\item{data}{Data.frame with raw moments.}

\item{Out}{What type of statistical moments do we want to get? 
Three types moments can be obtained: \code{"raw", "central", "normalized"}.}
}
\description{
Convert Raw Moments to Central or Normalized moments
}
\keyword{internal}
