% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBackTesting.R
\name{core.Accuracy}
\alias{core.Accuracy}
\title{Generate different types of output}
\usage{
core.Accuracy(u, u.hat, b, by = NULL, na.rm = TRUE)
}
\arguments{
\item{u}{Observed data}

\item{u.hat}{In-sample Forecast data}

\item{b}{Benchmark forecast data. Usualy a naive or rwd forecast.}

\item{by}{The accuracy measures are available in different output formats.
Option 1: \code{by = NULL} the agregated measure over ages and and years will 
be return. This is the default option. 
Option 2: \code{by = "x"} measure for all ages (agregated over time) are returned.
Option 3: \code{by = "y"} measures for all years (aggregated over ages) are returned.}

\item{na.rm}{A logical value indicating whether NA values should be stripped 
before the computation proceeds.}
}
\description{
Generate different types of output
}
\keyword{internal}
