% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_BBackTesting.R
\name{do.BBackTesting}
\alias{do.BBackTesting}
\title{Perform Out-of-Sample Testing of Mortality Forecasts Over Multiple Time Periods}
\usage{
do.BBackTesting(data, x, y, data.in = c("qx", "mx", "dx", "lx"), models,
  strategy = c(f = 20, h = 20, s = 2), level = 95,
  jumpchoice = c("actual", "fit"), verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{data.in}{Specify the type of input \code{data}. Various life table 
indices are accepted: \code{"qx", "mx", "dx", "lx"}.}

\item{models}{One or several mortality models to be estimated. 
The following options are available: \itemize{
  \item{\code{"MRW"}} -- The Multivariate Random-Walk (w/o Drift);
  \item{\code{"MRWD"}} -- The Multivariate Random-Walk with Drift;
  \item{\code{"LeeCarter"}} -- The Lee-Carter Mortality Model;
  \item{\code{"LiLee"}} -- The Li-Lee Mortality Model;
  \item{\code{"HyndmanUllah"}} -- The Hyndman-Ullah Mortality Model;
  \item{\code{"Oeppen"}} -- The Oeppen Mortality Model;
  \item{\code{"OeppenC"}} -- The Coherent Oeppen Mortality Model;
  \item{\code{"MEM2"}} -- The Maximum Entropy Mortality Model of order 2;
  \item{\code{"MEM3"}} -- The Maximum Entropy Mortality Model of order 3;
  \item{\code{"MEM4"}} -- The Maximum Entropy Mortality Model of order 4;
  \item{\code{"MEM5"}} -- The Maximum Entropy Mortality Model of order 5;
  \item{\code{"MEM6"}} -- The Maximum Entropy Mortality Model of order 6;
  \item{\code{"MEM7"}} -- The Maximum Entropy Mortality Model of order 7;
}}

\item{strategy}{Fitting-Forecasting strategy. Format: numerical vector.
The strategy \code{c(f, h, s)} consists in \code{f} number of years to use 
in fitting, \code{h} number of years to forecast and the step \code{s} to 
roll the evaluation window.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Perform Out-of-Sample Testing of Mortality Forecasts Over Multiple Time Periods
}
\examples{
x  <- 0:95
y  <- 1970:2016
dx <- HMD_male$dx$GBRTENW[paste(x), paste(y)]

BB <- do.BBackTesting(data = dx, x, y,
                      data.in = "dx", 
                      models = c("MRWD", "HyndmanUllah"),
                      strategy = c(20, 20, 3))

A <- evalAccuracy(BB, data.out = "ex")
A

R <- do.Ranking(A)
R
}
\seealso{
\code{\link{do.BackTesting}}
\code{\link{evalAccuracy.BBackTesting}}
\code{\link{evalRobustness.BBackTesting}}
}
\author{
Marius D. Pascariu
}
