% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_Ranking.R
\name{do.Ranking}
\alias{do.Ranking}
\title{Rank models based on the accuracy or robustness results}
\usage{
do.Ranking(A)
}
\arguments{
\item{A}{Table containing accuracy or robustness measures.}
}
\description{
Rank models based on the accuracy or robustness results
}
\examples{
# For examples go to ?do.BackTesting or ?do.BBackTesting
}
\seealso{
\code{\link{do.BackTesting}} 
\code{\link{do.BBackTesting}}
}
\author{
Marius D. Pascariu
}
