% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBackTesting.R
\name{doBackTesting}
\alias{doBackTesting}
\title{Perform in-sample testing of mortality forecasts}
\usage{
doBackTesting(data, x, y.fit, y.for, data.type, what, exogen, ci = 95,
  jumpchoice = "actual", measures = c("ME", "MAE", "MAPE", "sMAPE", "MRAE",
  "MASE"), ...)
}
\arguments{
\item{data}{Mortality data for fitting of the models and for validation; 
that is the data coresponding to the horizon to be predicted.}

\item{x}{Vector of ages.}

\item{y.fit}{Years to be considered in fitting}

\item{y.for}{Years to be forecast}

\item{data.type}{Type of data in 'data' argument. Options: 
\code{"qx", "mx", "dx", "lx"}.}

\item{what}{Specify the type of values to be extracted.}

\item{exogen}{Inlusion of exogenous variables.}

\item{ci}{Confidence level for prediction intervals.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{measures}{What accurracy measure to compute?}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Perform in-sample testing of mortality forecasts
}
