% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBBackTesting.R
\name{doBBackTesting}
\alias{doBBackTesting}
\title{Perform In-Sample Testing of Mortality Forecasts Over Multiple Time Periods}
\usage{
doBBackTesting(data, x, y, data.in = c("qx", "mx", "dx", "lx"), models,
  strategy = c(f = 20, h = 20, s = 2), level = 95,
  jumpchoice = c("actual", "fit"), verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{data.in}{Specify the type of input \code{data}. Various life table 
indices are accepted: \code{"qx", "mx", "dx", "lx"}.}

\item{models}{Mortality models to be evaluated.}

\item{strategy}{Fitting-Forecasting strategy. Format: numerical vector.
The strategy \code{c(f, h, s)} consists in \code{f} number of years to use in 
fitting, \code{h} number of years to forecast and the step \code{s} to roll the
evaluation window.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Perform In-Sample Testing of Mortality Forecasts Over Multiple Time Periods
}
\examples{
x  <- 0:95
y  <- 1970:2016
dx <- HMD_male$dx$GBRTENW[paste(x), paste(y)]

BB <- doBBackTesting(data = dx, x, y,
                     data.in = "dx", 
                     models = c("MRWD", "LeeCarter", "HyndmanUllah"),
                     strategy = c(20, 20, 3))

A <- evalAccuracy(BB, data.out = "ex")
A

R <- doRanking(A)
R
}
\seealso{
\code{\link{doBackTesting}}
\code{\link{evalAccuracy.doBBackTesting}}
\code{\link{evalRobustness.doBBackTesting}}
}
\author{
Marius D. Pascariu
}
