% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doForecasts.R
\name{doForecasts}
\alias{doForecasts}
\title{Forecast Human Mortality}
\usage{
doForecasts(object, h, level = 95, jumpchoice = c("actual", "fit"),
  ...)
}
\arguments{
\item{object}{An object of the class \code{MotalityModels} generated by the
\code{\link{doMortalityModels}} function.}

\item{h}{Number of years to be forecast in the future}

\item{level}{Confidence level for prediction intervals.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{...}{Further argumets to be passed to other methods.}
}
\description{
Forecast Human Mortality
}
\examples{
x = 0:100
y = 1985:1999
h = 17
D  <- MortalityForecast.data$dx[paste(x), paste(y)]
ex <- MortalityForecast.data$ex
exogen <- ex[paste(y)]

MM = c("LC", "FDM", "CoDa", "M6", "M6X")
M <- doMortalityModels(data = D, x, y, data.type = "dx",
                       models = MM, exogen = exogen)
                       
P <- doForecasts(M, h, level = 95, jumpchoice = "actual")

pex <- getForecasts(P, what = "ex")
}
