% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuracy.R
\name{doRanking}
\alias{doRanking}
\title{Rank models based on the accuracy results}
\usage{
doRanking(z)
}
\arguments{
\item{z}{Table containing accuracy measures.}
}
\description{
Rank models based on the accuracy results
}
