% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalAccuracy.R
\name{evalAccuracy.BBackTesting}
\alias{evalAccuracy.BBackTesting}
\title{Get Accuracy Measures from a \code{BBackTesting} object}
\usage{
\method{evalAccuracy}{BBackTesting}(object, data.out = c("qx", "mx",
  "dx", "lx", "Lx", "Tx", "ex"), measures = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{BBackTesting}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{measures}{What accuracy measure to compute? Various alternatives are 
available, \itemize{
 \item{Mean error measures: } \code{"ME", "MAE", "MAPE", "sMAPE", "sMRAE", "MASE"};
 \item{Median error measures: } \code{"MdE", "MdAE", "MdAPE", "sMdAPE", "sMdRAE", "MdASE"};
 \item{Squared error measures: } \code{"MSE", "RMSE", "RMSPE", "RMdSPE"};
 \item{Geometric mean measure for positive errors: } \code{"GMRAE"}.}
 If \code{measures = NULL} all the measures will be computed.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Get Accuracy Measures from a \code{BBackTesting} object
}
\details{
See \insertCite{hyndman2006;textual}{MortalityForecast} for a 
comprehensive discussion of the accuracy measures.
}
\examples{
# For examples go to ?do.BBackTesting
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{do.BBackTesting}}
}
\author{
Marius D. Pascariu
}
