% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalAccuracy.R
\name{evalAccuracy}
\alias{evalAccuracy}
\title{Generic for Computing Accuracy Measure from the \code{doBackTesting} and 
\code{doBBackTesting} objects}
\usage{
evalAccuracy(object, ...)
}
\arguments{
\item{object}{An object of the class \code{doBackTesting} or \code{doBBackTesting}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Generic for Computing Accuracy Measure from the \code{doBackTesting} and 
\code{doBBackTesting} objects
}
\examples{
# For examples go to ?doBackTesting and ?doBBackTesting
}
\author{
Marius D. Pascariu
}
\keyword{internal}
