% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalRobustness.R
\name{evalRobustness.BBackTesting}
\alias{evalRobustness.BBackTesting}
\title{Evaluating the Robustness of the Forcasting Mortality Models}
\usage{
\method{evalRobustness}{BBackTesting}(object, data.out = c("qx", "mx",
  "dx", "lx", "Lx", "Tx", "ex"), measures = c("MD", "MAD", "sMRAD",
  "MASD"), ...)
}
\arguments{
\item{object}{An object of class \code{BBackTesting}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{measures}{What accurracy measure to compute? Various alternatives are 
available, \itemize{
 \item{Mean error measures: } \code{"ME", "MAE", "MAPE", "sMAPE", "sMRAE", "MASE"};
 \item{Median error measures: } \code{"MdE", "MdAE", "MdAPE", "sMdAPE", "sMdRAE", "MdASE"};
 \item{Squared error measures: } \code{"MSE", "RMSE", "RMSPE", "RMdSPE"};
 \item{Geometric mean measure for positive errors: } \code{"GMRAE"}.}
 If \code{measures = NULL} all the measures will be computed.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Evaluating the Robustness of the Forcasting Mortality Models
}
\examples{
# For examples go to ?do.BBackTesting
}
\seealso{
\code{\link{do.BBackTesting}}
\code{\link{evalAccuracy.BBackTesting}}
}
\author{
Marius D. Pascariu
}
