% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalRobustness.R
\name{evalRobustness.doBBackTesting}
\alias{evalRobustness.doBBackTesting}
\title{Evaluating the Robustness of the Forcasting Mortality Models}
\usage{
\method{evalRobustness}{doBBackTesting}(object, data.out = c("qx", "mx",
  "dx", "lx", "Lx", "Tx", "ex"), measures = c("MD", "MAD", "sMRAD"), ...)
}
\arguments{
\item{object}{An object of class \code{doBBackTesting}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{measures}{What accurracy measure to compute? Various alternatives are 
available, \itemize{
 \item{Mean error measures: } \code{"ME", "MAE", "MAPE", "sMAPE", "sMRAE", "MASE"};
 \item{Median error measures: } \code{"MdE", "MdAE", "MdAPE", "sMdAPE", "sMdRAE", "MdASE"};
 \item{Squared error measures: } \code{"MSE", "RMSE", "RMSPE", "RMdSPE"};
 \item{Geometric mean measure for positive errors: } \code{"GMRAE"}.}}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Evaluating the Robustness of the Forcasting Mortality Models
}
\examples{
# For examples go to ?doBBackTesting
}
\seealso{
\code{\link{doBBackTesting}}
\code{\link{evalAccuracy.doBBackTesting}}
}
\author{
Marius D. Pascariu
}
