% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalRobustness.R
\name{evalRobustness}
\alias{evalRobustness}
\title{Generic for evaluating the robustness of the mortality models
tested with the \code{do.BBackTesting} function.}
\usage{
evalRobustness(object, ...)
}
\arguments{
\item{object}{An object of the class \code{BBackTesting};}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Generic for evaluating the robustness of the mortality models
tested with the \code{do.BBackTesting} function.
}
\examples{
# For examples go to ?do.BBackTesting
}
\author{
Marius D. Pascariu
}
\keyword{internal}
