% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StMoMo-models.R
\name{FDM}
\alias{FDM}
\title{Functional Demographic Model}
\usage{
FDM(data, x, y, ...)
}
\arguments{
\item{data}{Mortality data}

\item{x}{Vector of ages.}

\item{y}{Vector of years.}
}
\description{
Functional Demographic Model
}
\seealso{
\code{\link[demography]{fdm}}
}
\keyword{internal}
