% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_fit.R
\name{find_ages_and_years}
\alias{find_ages_and_years}
\title{Find age and year vectors}
\usage{
find_ages_and_years(data, x, y)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data (dx) 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages to be considered in the 
fitting procedure. If \code{NULL} the number fitted data points is equal 
with the observed number of data points.
This argument can be used for example to estimate a density between 0 and 130 
given the fact that the input dataset contains values for 0-100 only.}

\item{y}{Numerical vector indicating the years in \code{data}. 
Default: \code{NULL}.}
}
\description{
Find age and year vectors
}
\keyword{internal}
