% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_arima}
\alias{find_arima}
\title{Find ARIMA order of a time series}
\usage{
find_arima(a)
}
\arguments{
\item{a}{A numerical vector.}
}
\description{
Find ARIMA order of a time series
}
\keyword{intermal}
