% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{find_grob_cells}
\alias{find_grob_cells}
\title{Utility funtion for plot.getAccuracy}
\usage{
find_grob_cells(tg, object)
}
\arguments{
\item{tg}{An object of the class \code{\link{tableGrob}}.}

\item{object}{An object of the class \code{\link{getAccuracy}}.}
}
\description{
Utility funtion for plot.getAccuracy
}
\keyword{internal}
