% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDensity.R
\name{findDensity}
\alias{findDensity}
\title{Estimate density function based on Maximum Entropy method (MaxEnt)}
\usage{
findDensity(data, x, omega = 110, verbose = FALSE)
}
\arguments{
\item{data}{A numeric vector or a data.frame/matrix with raw moments.}

\item{x}{A numeric vector with ages. The density frequencies are 
returned only for the specified ages.}

\item{omega}{The maximum age to be used in MaxEnt estimation. The 
algorithm is reliable and robust when the underlying distribution is closing 
on the right-hand side. In the case of age-at-death distribution, \code{omega} 
should be an old age; otherwise the MaxEnt might fail to converge. 
Numeric scalar.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}
}
\value{
An object with the estimated densities.
}
\description{
Function for estimating the probability density function starting from 
a limited number of raw statistical moments. The method is based on 
Maximum Entropy method \insertCite{mead1984}{MortalityForecast}.
}
\examples{
# Example 1 -- simple case ---------------

x <- 0:120
# Raw moments M0 - M7
raw_moments <- c(1, 68.75099, 4991.724, 371531.9, 28199680, 
                 2176435499, 170477697491, 1.353288e+13)
ed <- findDensity(raw_moments, x)

# Example 2 -- estimat for entire table --

x   <- 0:110
y   <- 1965:2016
dx  <- HMD_male$dx$GBRTENW[paste(x), paste(y)]
mom <- findMoments(dx, x, y, n = 7)$raw.moments

dx.hat <- findDensity(mom, x)

# Visual check --
yr = "1990"
observed.dx <- dx[, yr]/sum(dx[, yr])
estimated.dx <- dx.hat$density[, yr]
plot(observed.dx, pch = 16, main = yr)
lines(estimated.dx, col = 2)
legend("topleft", legend = c("observed", "estimated"), col = 1:2, 
       pch = c(16, NA), lty = c(NA, 1), lwd = 2, bty = "n")
}
\references{
\insertAllCited{}
}
\author{
Adam Lenart and Marius D. Pascariu.
}
