% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMoments.R
\name{findMoments}
\alias{findMoments}
\title{Compute Statistical Moments}
\usage{
findMoments(data, x, y = NULL, n, na.rm = TRUE)
}
\arguments{
\item{data}{A data.frame or matrix containing 3 columns: Year, Age, Dx.}

\item{x}{Vector of ages.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{n}{The maximum order of the moments to be computed. 
The order should be at least 2.}

\item{na.rm}{Logical value. If \code{TRUE}, remove \code{NA} values. 
Otherwise, keep \code{NA} values.}
}
\value{
An object containing:
\itemize{
\item \code{central.moments} --- Moments about the mean (c = mean). The zeroth moment is 
the total probability (i.e. one), the first moment is zero, 
the second central moment is the variance, the third central moment 
is the skewness (with normalization), and the fourth central moment 
is the kurtosis (with normalization).
\item \code{raw.moments} --- Moments about zero (c = 0). The first moment is the mean.
\item \code{normalized.moments} --- Normalized moments. 
}
}
\description{
Compute raw and central statistical moments of a distribution of deaths.
}
\details{
If \eqn{f(x)} is a probability density function, then \eqn{\mu(n)} is called 
the n-th moment of the probability distribution, where: \cr
\deqn{\mu_{n} = \int_{-\infty}^{\infty} (x-c)^n f(x)dx.}{\mu(n) = \int (x-c)^n f(x)dx.}
}
\examples{
x  <- 0:110
y  <- 1960:2016
dx <- MortalityForecast.data$dx
findMoments(data = dx, x = x, y = y, n = 4)
}
\seealso{
\code{\link{convertMoments}}
\code{\link[moments]{all.moments}} 
\code{\link[moments]{raw2central}}
}
