% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{fit_LeeCarter}
\alias{fit_LeeCarter}
\title{The Lee-Carter Mortality Model}
\usage{
fit_LeeCarter(data, x, y, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
Object of class \dQuote{lca} with the following components:
\item{label}{Name of region}
\item{age}{Ages from \code{data} object.}
\item{year}{Years from \code{data} object.}
\item{<series>}{Matrix of mortality or fertility data as contained in \code{data}. It takes the name given by the series argument.}
\item{ax}{Average deathrates across fitting period}
\item{bx}{First principal component in Lee-Carter model}
\item{kt}{Coefficient of first principal component}
\item{residuals}{Functional time series of residuals.}
\item{fitted}{Functional time series containing estimated mortality or fertility rates from model}
\item{varprop}{Proportion of variance explained by model.}
\item{y}{The data stored as a functional time series object.}
\item{mdev}{Mean deviance of total and base lack of fit, as described in Booth, Maindonald and Smith.}
}
\description{
The Lee-Carter Mortality Model
}
\details{
\insertNoCite{lee1992}{MortalityForecast}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fit_HyndmanUllah}}
\code{\link{fit_Oeppen}}
}
