% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter2.R
\name{fit_LeeCarter2}
\alias{fit_LeeCarter2}
\title{The Lee-Carter Mortality Model}
\usage{
fit_LeeCarter2(data, x = NULL, y = NULL, verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Fit the Lee-Carter mortality model
}
\details{
\insertNoCite{lee1992}{MortalityForecast}
}
\examples{
# Data
x  <- 0:89
y  <- 1985:2014
mx <- MortalityForecast.data$mx[paste(x), paste(y)]

# Fit the model
M <- fit_LeeCarter2(data = mx, x = x, y = y)
M
summary(M)

# Check residuals
R <- residuals(M)

plot(R, plotType = "scatter")
plot(R, plotType = "colourmap")
plot(R, plotType = "signplot")

# Forecast 
P <- predict(M, h = 20)
P
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{predict.LeeCarter2}}
}
