% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-models.R
\name{fitHyndmanUllah}
\alias{fitHyndmanUllah}
\title{The Functional Demographic Model}
\usage{
fitHyndmanUllah(data, x, y, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
Object of class \dQuote{fdm} with the following components:
\item{label}{Name of country}
\item{age}{Ages from \code{data} object.}
\item{year}{Years from \code{data} object.}
\item{<series>}{Matrix of demographic data as contained in \code{data}. It takes the name given by the series argument.}
\item{fitted}{Matrix of fitted values.}
\item{residuals}{Residuals (difference between observed and fitted).}
\item{basis}{Matrix of basis functions evaluated at each age level (one column for each basis function). The first column is the fitted mean.}
\item{coeffs}{Matrix of coefficients (one column for each coefficient series). The first column are all ones.}
\item{mean.se}{Standard errors for the estimated mean function.}
\item{varprop}{Proportion of variation explained by each basis function.}
\item{weights}{Weight associated with each time period.}
\item{v}{Measure of variation for each time period.}
\item{type}{Data type (mortality, fertility, etc.)}
\item{y}{The data stored as a functional time series object.}
}
\description{
The Functional Demographic Model
}
\details{
\insertNoCite{hyndman2007}{MortalityForecast}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fitLeeCarter}}
\code{\link{fitOeppen}}
}
