% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-models.R
\name{fitLeeCarter}
\alias{fitLeeCarter}
\title{Lee-Carter Demographic Model}
\usage{
fitLeeCarter(data, x, y, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
Object of class \dQuote{lca} with the following components:
\item{label}{Name of region}
\item{age}{Ages from \code{data} object.}
\item{year}{Years from \code{data} object.}
\item{<series>}{Matrix of mortality or fertility data as contained in \code{data}. It takes the name given by the series argument.}
\item{ax}{Average deathrates across fitting period}
\item{bx}{First principal component in Lee-Carter model}
\item{kt}{Coefficient of first principal component}
\item{residuals}{Functional time series of residuals.}
\item{fitted}{Functional time series containing estimated mortality or fertility rates from model}
\item{varprop}{Proportion of variance explained by model.}
\item{y}{The data stored as a functional time series object.}
\item{mdev}{Mean deviance of total and base lack of fit, as described in Booth, Maindonald and Smith.}
}
\description{
Lee-Carter Demographic Model
}
\details{
All mortality or fertility data are assumed to be in matrices of
mortality or fertility rates within \code{data$rate}. Each row is one age group
(assumed to be single years). Each column is one year. The
function produces a model for the \code{series} mortality or fertility rate matrix
within \code{data$rate}. Forecasts from this model can be obtained using \code{\link{forecast.lca}}.
}
\seealso{
\code{\link{fitHyndmanUllah}}
\code{\link{coda}}
}
