% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_fit.R
\name{fitMaxEntMortality}
\alias{fitMaxEntMortality}
\title{The Maximum-Entropy Mortality Model}
\usage{
fitMaxEntMortality(data, x = NULL, y = NULL, n = 5,
  verbose = FALSE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{n}{The maximum order of the moments to be used.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
The output is an object of class \code{"fitMaxEntMortality"} with 
the components:
 \item{input}{List with arguments provided in input. Saved for convenience.}
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given arguments;}
 \item{coefficients}{Estimated coefficients;}
 \item{fitted.values}{Fitted values of the estimated model;}
 \item{observed.values}{Observed values used in fitting;}
 \item{fitted.raw.moments}{Fitted raw moments;}
 \item{observed.raw.moments}{Observed raw moments of the input data;}
 \item{residuals}{Deviance residuals;} 
 \item{VAR}{Object containing the components of the fitted time 
 series model to the extrapolate moments;}
 \item{x}{Vector of ages used in the fitting;} 
 \item{y}{Vector of years used in the fitting.}
}
\description{
The Maximum-Entropy Mortality Model
}
\examples{
# Data
x  <- 0:110
y  <- 1985:2014
dx <- MortalityForecast.data$dx[paste(x), paste(y)]

# Fit model the Maximum-Entropy Mortality of order 5,
# that is using the first 6 statistical moments (moment 0 to 5).

M <- fitMaxEntMortality(data = dx, x = x, y = y, n = 5)
M
R <- residuals(M)

plot(M, plotType = "observed")
plot(M, plotType = "fitted")

plot(R, plotType = "scatter")
plot(R, plotType = "colourmap")
plot(R, plotType = "signplot")
}
\seealso{
\code{\link{predict.fitMaxEntMortality}} 
\code{\link{plot.fitMaxEntMortality}} 
\code{\link{findMoments}}
}
