% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRW_fit.R
\name{fitMRandomWalk}
\alias{fitMRandomWalk}
\title{Multivariate Random Walk Model}
\source{
The original implementation of this function was taken from StMoMo R package.
}
\usage{
fitMRandomWalk(data, x = NULL, y = NULL, include.drift = TRUE, ...)
}
\arguments{
\item{data}{Numeric matrix with a multivariate time series.
Series are arranged in rows with columns representing time.}

\item{x}{Numerical vector indicating the age classes in input \code{data}. 
Optional. This is used to label the output tables and related plots.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}.
Optional. This is used to label the output tables and related plots.
Default: \code{NULL}.}

\item{include.drift}{Should the Random Walk model include a linear drift term? 
Default: \code{TRUE}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
An object of class \code{"fitMRandomWalk"} with components:
 \item{input}{ A list with the input data.}
 \item{drift}{ A vector with the estimated drift.}
 \item{sigma}{ A matrix with the estimated variance covariance matrix.}
 \item{fitted}{ Fitted values.}
 \item{residuals}{ Residuals from the fitted model. That is 
 observed minus fitted values.}
}
\description{
Fit a Multivariate Random Walk Model to the input \code{data}, a
multivariate time series.
}
\details{
For further information on the Multivariate Random Walk with
drift see Appendix B in \insertCite{haberman2011;textual}{MortalityForecast}.
}
\examples{
# Forecast mortality using a Multivariate Random Walk with Drift model
x  <- 0:100
y  <- 1980:2000
mx <- MortalityForecast.data$mx[paste(x), paste(y)]

M <- fitMRandomWalk(data = log(mx), x = x, y = y, include.drift = TRUE)
P <- predict(M, h = 16)
R <- residuals(M)

pv <- exp(P$predicted.values)
matplot(pv, type = "l", log = "y")
matplot(t(pv), type = "l", log = "y")
}
\references{
\insertAllCited{}
}
\author{
Marius D. Pascariu
}
