% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODA_fit.R
\name{fitOeppen.input.check}
\alias{fitOeppen.input.check}
\title{Validate input values}
\usage{
fitOeppen.input.check(X)
}
\arguments{
\item{X}{A list with input arguments provided in \code{\link{fitOeppen}} function}
}
\description{
Validate input values
}
\keyword{internal}
