% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMortality.R
\name{fitted.MortalityModels}
\alias{fitted.MortalityModels}
\title{Get Fitted values}
\usage{
\method{fitted}{MortalityModels}(object, type = c("qx", "mx", "dx", "lx",
  "Lx", "Tx", "ex"), ...)
}
\arguments{
\item{object}{An object of the class \code{MortalityModels}.}

\item{type}{Specify the type of fitted values to be extracted.}

\item{...}{Further argumets to be passed to other methods.}
}
\description{
Get Fitted values
}
