% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MEM_plot.R
\name{fitted2dens}
\alias{fitted2dens}
\title{Prepare data for ggplots in plot.MEM function}
\usage{
fitted2dens(mat, x, y, ny = 7, lx0 = 300)
}
\arguments{
\item{mat}{Matrix containing the observed or fitted value.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}

\item{lx0}{Adjusting parameter.}
}
\description{
Prepare data for ggplots in plot.MEM function
}
\examples{
x  <- 0:110
y  <- 1965:2014
dx <- HMD_male$dx$GBRTENW[paste(x), paste(y)]
M  <- model_MEM(dx, x, y, n = 5)
fitted2dens(fitted(M), x, y[-1])
}
\keyword{internal}
