% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Observed.R
\name{get.Observed}
\alias{get.Observed}
\title{Get Observed Values}
\usage{
get.Observed(object = NULL, data.out = c("qx", "mx", "dx", "lx", "Lx",
  "Tx", "ex"), data.in = NULL, data = NULL, x = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{MortalityModels}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{data.in}{Specify the type of input \code{data}. Various life table 
indices are accepted: \code{"qx", "mx", "dx", "lx"}.}

\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Extract the values from an object containing an estimated mortality model
(e.g. \code{LeeCarter, Oeppen} etc.). This function allows the user to read 
the values already converted to a life table index (e.g. life expectancy) 
by specifying the requiered format in the argument \code{data.out}.
}
\examples{
# For examples go to ?do.MortalityModels
}
\seealso{
\code{\link{do.MortalityModels}}
}
\author{
Marius D. Pascariu
}
