% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Residuals.R
\name{get.Residuals}
\alias{get.Residuals}
\title{Get Deviance Residuals}
\usage{
get.Residuals(object, data.out = c("qx", "mx", "dx", "lx", "Lx", "Tx",
  "ex"), ...)
}
\arguments{
\item{object}{An object of class \code{MortalityModels}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Extract the values from an object containing an estimated mortality model
(e.g. \code{LeeCarter, Oeppen} etc.). This function allows the user to read 
the values already converted to a life table index (e.g. life expectancy) 
by specifying the requiered format in the argument \code{data.out}.
}
\examples{
# For examples go to ?do.MortalityModels
}
\seealso{
\code{\link{do.MortalityModels}}
}
\author{
Marius D. Pascariu
}
