% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{get_dx_values}
\alias{get_dx_values}
\title{Internal function}
\usage{
get_dx_values(object, kt, jumpchoice, y)
}
\arguments{
\item{object}{An object of class \code{Oeppen}.}

\item{kt}{Estimated kt vector of parameters}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{ax}{Estimated ax vector of parameters}

\item{bx}{Estimated bx vector of parameters}

\item{fit}{Fitted values}
}
\description{
Internal function
}
\keyword{internal}
