% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{get_mx_values}
\alias{get_mx_values}
\title{Get m[x] values and confidence intervals based on k[t] forecast
In this function we compute the m[x] values based on the extrapolation of
the k[t] time-series. If necessary an adjustment for the jump-off is 
provided.}
\usage{
get_mx_values(object, jumpchoice, y, kt, B.kt = NULL)
}
\arguments{
\item{object}{An object of class \code{LeeCarter}.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{y}{Numerical vector indicating the years in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{kt}{Predicted k[t] values in the model;}

\item{B.kt}{Predicted k[t] values of the benchmark model, used in the Li-Lee model only.}
}
\description{
Get m[x] values and confidence intervals based on k[t] forecast
In this function we compute the m[x] values based on the extrapolation of
the k[t] time-series. If necessary an adjustment for the jump-off is 
provided.
}
\keyword{internal}
