% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuracy.R
\name{getAccuracy.doBackTesting}
\alias{getAccuracy.doBackTesting}
\title{Get Accuracy Measures from \code{doBackTesting}}
\usage{
\method{getAccuracy}{doBackTesting}(object, data.out = c("qx", "mx",
  "dx", "lx", "Lx", "Tx", "ex"), measures = c("ME", "MAE", "MAPE",
  "sMAPE", "sMRAE", "MASE"), ...)
}
\arguments{
\item{object}{An object of class \code{doBackTesting}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{measures}{What accurracy measure to compute? Various alternatives are 
available, \itemize{
 \item{Mean error measures: } \code{"ME", "MAE", "MAPE", "sMAPE", "sMRAE", "MASE"};
 \item{Median error measures: } \code{"MdE", "MdAE", "MdAPE", "sMdAPE", "sMdRAE", "MdASE"};
 \item{Squared error measures: } \code{"MSE", "RMSE", "RMSPE", "RMdSPE"};
 \item{Geometric mean measure for positive errors: } \code{"GMRAE"}.}}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Get Accuracy Measures from \code{doBackTesting}
}
\details{
See \insertCite{hyndman2006;textual}{MortalityForecast} for a 
comprehensive discussion of the accuracy measures.
}
\examples{
# For examples go to ?doBackTesting
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{doBackTesting}}
}
\author{
Marius D. Pascariu
}
