% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuracy.R
\name{getAccuracy}
\alias{getAccuracy}
\title{Get Accuracy Measures}
\usage{
getAccuracy(object, data, x, y, data.type, what, measures, na.rm, ...)
}
\arguments{
\item{object}{An object of the class \code{doForecasts}.}

\item{data}{Validation set of demographic data.}

\item{x}{Vector of ages.}

\item{y}{Vector of years.}

\item{data.type}{Type of data in 'data' argument. Options: 
\code{"qx", "mx", "dx", "lx"}.}

\item{what}{Specify the type of values to be extracted.}

\item{measures}{What accurracy measure to compute?}

\item{na.rm}{A logical value indicating whether NA values should be stripped 
before the computation proceeds.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Get Accuracy Measures
}
\examples{
x = 0:100
y1 = 1980:1995
y2 = 1996:2005
h = max(y2) - max(y1)

D1 <- dxForecast::dxForecast.data$dx$male[paste(x), paste(y1)]
D2 <- dxForecast::dxForecast.data$dx$male[paste(x), paste(y2)]

MM = c("LC", "FDM", "CoDa", "M6")
M <- doMortalityModels(data = D1, x, y1, data.type = "dx", models = MM)
P <- doForecasts(M, h)
A <- getAccuracy(P, D2, x = 0:95, data.type = "dx", what = "qx")

A
plot(A)
}
