% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMortalityModels.R
\name{getFitted}
\alias{getFitted}
\title{Get Fitted Values}
\usage{
getFitted(object, what = c("qx", "mx", "dx", "lx", "Lx", "Tx", "ex"),
  ...)
}
\arguments{
\item{object}{An object of the class \code{MortalityModels}.}

\item{what}{Specify the type of values to be extracted.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Get Fitted Values
}
