% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMortality.R
\name{getMortalityModels}
\alias{getMortalityModels}
\title{Wrapper for function dealing with mortality models}
\usage{
getMortalityModels(data, x, y, data.type = c("qx", "mx", "dx"), ...)
}
\arguments{
\item{data}{Mortality data}

\item{x}{Vector of ages.}

\item{y}{Vector of years.}

\item{data.type}{Type of data in 'data' argument. Options: 
\code{"qx", "mx", "dx"}.}

\item{...}{Further argumets to be passed to other methods.}
}
\description{
Wrapper for function dealing with mortality models
}
\examples{
x = 0:110
y = 1960:2016
dxm <- dxForecast::dxForecast.data$dx$male[paste(x), paste(y)]
dx  <- apply(dxm, 2, function(x) x/sum(x))

M <- getMortalityModels(data = dx, x, y, data.type = "dx")
P <- predict(M, h = 16, ci = 95, jumpchoice = "actual")


fmx <- fitted(M, type = "dx")
fex <- fitted(M, type = "ex")

pex <- predicted(P, type = "ex") 
}
