% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_Plot_fit.R
\name{ggplotDistribConvergence}
\alias{ggplotDistribConvergence}
\title{Plot Convergence in Age at Death Distribution}
\usage{
ggplotDistribConvergence(mat, x, ny, quant)
}
\arguments{
\item{mat}{Matrix containing the observed or fitted value.}

\item{x}{Numerical vector indicating the ages to be considered in the 
fitting procedure. If \code{NULL} the number fitted data points is equal 
with the observed number of data points.
This argument can be used for example to estimate a density between 0 and 130 
given the fact that the input dataset contains values for 0-100 only.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}
}
\description{
Plot Convergence in Age at Death Distribution
}
\keyword{internal}
