% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StMoMo-models.R
\name{LC}
\alias{LC}
\title{Lee-Carter Mortality Model as implemented in StMoMo package}
\usage{
LC(data, x, y, link = "log", lx0 = 1e+05, verbose = FALSE)
}
\arguments{
\item{data}{Mortality data}

\item{x}{Vector of ages.}

\item{y}{Vector of years.}

\item{link}{defines the link function and random component associated with 
the mortality model. \code{"log"} would assume that deaths follow a 
Poisson distribution and use a log link while \code{"logit"} would assume 
that deaths follow a Binomial distribution and a logit link.}

\item{lx0}{lx0}
}
\description{
Lee-Carter Mortality Model as implemented in StMoMo package
}
\keyword{internal}
