% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{model.LeeCarter}
\alias{model.LeeCarter}
\title{The Lee-Carter Mortality Model}
\usage{
model.LeeCarter(data, x = NULL, y = NULL, verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
The output is a list with the components:
 \item{input}{List with arguments provided in input. Saved for convenience;}
 \item{info}{Short details about the model;}
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given 
 arguments;}
 \item{coefficients}{Estimated coefficients;}
 \item{fitted.values}{Fitted values of the estimated model;}
 \item{observed.values}{The observed values used in fitting arranged in the 
 same format as the fitted.values;}
 \item{residuals}{Deviance residuals;} 
 \item{x}{Vector of ages used in the fitting;} 
 \item{y}{Vector of years used in the fitting.}
}
\description{
Fit the Lee-Carter mortality model
}
\details{
\insertNoCite{lee1992}{MortalityForecast}
}
\examples{
# Data
x  <- 0:89
y  <- 1985:2014
mx <- HMD_male$mx$GBRTENW[paste(x), paste(y)]

# Fit the model
M <- model.LeeCarter(data = mx, x = x, y = y)
M
summary(M)

# Check residuals
R <- residuals(M)

plot(R, plotType = "scatter")
plot(R, plotType = "colourmap")
plot(R, plotType = "signplot")

# Forecast 
P <- predict(M, h = 20)
P
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{predict.LeeCarter}}
}
