% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MEM.R
\name{model.MEM}
\alias{model.MEM}
\title{The Maximum-Entropy Mortality Model}
\usage{
model.MEM(data, x = NULL, y = NULL, n = 5, verbose = FALSE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{n}{The maximum order of the moments to be used.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
The output is an object of class \code{MEM} with 
the components:
 \item{input}{List with arguments provided in input. Saved for convenience;}
 \item{info}{Short details about the model;}
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given 
 arguments;}
 \item{coefficients}{Estimated coefficients;}
 \item{fitted.values}{Fitted values of the estimated model;}
 \item{observed.values}{Observed values used in fitting;}
 \item{fitted.raw.moments}{Fitted raw moments;}
 \item{observed.raw.moments}{Observed raw moments of the input data;}
 \item{residuals}{Deviance residuals;} 
 \item{random.walk.model}{Object containing the components of the fitted time 
 series model to the extrapolate moments;}
 \item{x}{Vector of ages used in the fitting;} 
 \item{y}{Vector of years used in the fitting.}
}
\description{
The Maximum-Entropy Mortality Model
}
\examples{
# Data
x  <- 0:110
y  <- 1985:2014
dx <- HMD_male$dx$GBRTENW[paste(x), paste(y)]

# Fit model the Maximum-Entropy Mortality of order 5,
# that is using the first 6 statistical moments (moment 0 to 5).

M <- model.MEM(data = dx, x = x, y = y, n = 5)
M
R <- residuals(M)

plot(M, plotType = "observed")
plot(M, plotType = "fitted")

plot(R, plotType = "scatter")
plot(R, plotType = "colourmap")
plot(R, plotType = "signplot")

# Perform forecasts
P <- predict(M, h = 16, x.h = 0:120)

plot(P, plotType = "mean")
plot(P, plotType = "lower")
plot(P, plotType = "upper")

plot(P, M, plotType = "raw_moments")
plot(P, M, plotType = "normalised_moments")
}
\seealso{
\code{\link{predict.MEM}} 
\code{\link{plot.MEM}} 
\code{\link{find.moments}}
}
