% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MRW.R
\name{model.MRW}
\alias{model.MRW}
\title{The Multivariate Random Walk Model}
\source{
The original implementation of this function was taken from 
\code{StMoMo} R package.
}
\usage{
model.MRW(data, x = NULL, y = NULL, include.drift = TRUE,
  lowess.smooth = TRUE, ...)
}
\arguments{
\item{data}{Numeric matrix with a multivariate time series.
Series are arranged in rows with columns representing time.}

\item{x}{Numerical vector indicating the age classes in input \code{data}. 
Optional. This is used to label the output tables and related plots.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}.
Optional. This is used to label the output tables and related plots.
Default: \code{NULL}.}

\item{include.drift}{Should the Random Walk model include a linear drift 
term? Default: \code{TRUE}.}

\item{lowess.smooth}{Logical. Should the estimatated vector of drift 
parameters be smoothed using a lowess function? Default: \code{TRUE}. 
This helps avoiding highly divergent trends between time-series
when the model is used to do predictions.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
An object of class \code{MRW} with components:
 \item{input}{A list with the input data;}
 \item{info}{Short details about the model;}
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given 
 arguments;}
 \item{coefficients}{A vector with the estimated drift parameters;}
 \item{fitted.values}{Fitted values of the estimated model;}
 \item{observed.values}{The observed values used in fitting arranged in the 
 same format as the fitted.values;}
 \item{residuals}{Residuals from the fitted model. That is 
 observed minus fitted values;}
 \item{sigma}{A matrix with the estimated variance covariance matrix;}
 \item{x}{Vector of ages used in fitting;} 
 \item{y}{Vector of years used in fitting.}
}
\description{
Fit a Multivariate Random Walk Model to the input \code{data}, a
multivariate time series.
}
\details{
For further information on the Multivariate Random Walk with
drift see Appendix B in \insertCite{haberman2011;textual}{MortalityForecast}.
}
\examples{
# Forecast mortality using a Multivariate Random Walk with Drift model
x  <- 0:100
y  <- 1980:2000
mx <- HMD_male$mx$GBRTENW[paste(x), paste(y)]

M <- model.MRW(data = log(mx), x = x, y = y, include.drift = TRUE)
P <- predict(M, h = 16)
R <- residuals(M)

pv <- exp(P$predicted.values)
matplot(pv, type = "l", log = "y")
matplot(t(pv), type = "l", log = "y")
}
\references{
\insertAllCited{}
}
\author{
Marius D. Pascariu
}
