% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{model.Oeppen}
\alias{model.Oeppen}
\title{The Oeppen Mortality Model (Oeppen -- CoDa)}
\usage{
model.Oeppen(data, x = NULL, y = NULL, verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
The output is a list with the components:
 \item{input}{List with arguments provided in input. Saved for convenience;}
 \item{info}{Short details about the model;}
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given 
 arguments;}
 \item{coefficients}{Estimated coefficients;}
 \item{fitted.values}{Fitted values of the estimated model;}
 \item{observed.values}{The observed values used in fitting arranged in the 
 same format as the fitted.values;}
 \item{residuals}{Deviance residuals;} 
 \item{x}{Vector of ages used in the fitting;} 
 \item{y}{Vector of years used in the fitting.}
}
\description{
Fit the Oeppen model for forecasting the life table 
distribution of deaths. This is a Lee-Carter type model adapted to a 
compositional-data framework (CoDa). A key difference 
between the \insertCite{lee1992;textual}{MortalityForecast}
method and the Oeppen model is that the former fits and 
forecasts the death rates (mx) while the latter is based on the life table 
death distribution (dx). 
\insertCite{@See @oeppen2008 and @bergeron2017;textual}{MortalityForecast} 
for a detail description and mathematical formulation.
}
\examples{
# Example 1 ----------------------
# Data
x  <- 0:100
y  <- 1980:2016
dx <- HMD_male$dx$GBRTENW[paste(x), paste(y)]

# If data contains zero's we have to replace them with very small 
# values in order to avoid errors in fitting. replace.zeros() will do it.
dx <- replace.zeros(dx)

# Fit model
M <- model.Oeppen(data = dx, x = x, y = y)
M

summary(M)
coef(M)

# Plot observed and fitted values
plot(M, plotType = "observed")
plot(M, plotType = "fitted")

# Plot residuals
R <- residuals(M)
plot(R, plotType = "scatter")
plot(R, plotType = "colourmap")
plot(R, plotType = "signplot")

# Perform forecasts
P  <- predict(M, h = 16)
P

plot(P, plotType = "mean")
plot(P, plotType = "lower")
plot(P, plotType = "upper")

#' # Example 2 ----------------------
# One can specify manually the ARIMA order, a drift to be included or not, 
# and the jump choice of the first forecast year.
P2 <- predict(M, h = 20, 
              order = c(0,1,1), 
              include.drift = FALSE, 
              jumpchoice = "fit")

\dontrun{
# Example 3 ----------------------
# Compute life tables using forecast values using the MortalityLaws R package
library(MortalityLaws)
dx <- P$predicted.values
lt <- LifeTable(x = P$x, dx = dx)
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{predict.Oeppen}}
\code{\link{plot.Oeppen}}
}
\author{
Marius D. Pascariu and Marie-Pier Bergeron-Boucher
}
