% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_OeppenC.R
\name{model.OeppenC}
\alias{model.OeppenC}
\title{The Coherent Oeppen Mortality Model (Oeppen-C)}
\usage{
model.OeppenC(data, data.B, x = NULL, y = NULL, verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{data.B}{A data.frame or a matrix containing mortality data for the 
benchmark population. Must be the same format as in \code{data};}

\item{x}{Numerical vector indicating the ages in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
The output is a list with the components:
 \item{input}{List with arguments provided in input. Saved for convenience;}
 \item{info}{Short details about the model;}
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given 
 arguments;}
 \item{coefficients}{Estimated coefficients;}
 \item{fitted.values}{Fitted values of the estimated model;}
 \item{observed.values}{The observed values used in fitting arranged in the 
 same format as the fitted.values;}
 \item{residuals}{Deviance residuals;} 
 \item{x}{Vector of ages used in the fitting;} 
 \item{y}{Vector of years used in the fitting;} 
 \item{benchmark}{An object of class \code{LeeCarter} containing the fitted
 model for the benchmark population.}
}
\description{
The Coherent Oeppen Mortality Model (Oeppen-C)
}
\examples{
# Example 1 ----------------------
# Data
x  <- 0:110
y  <- 1980:2016
dx_M <- HMD_male$dx$USA[paste(x), paste(y)]
dx_F <- HMD_female$dx$USA[paste(x), paste(y)]

# Replace zeros 
dx_M <- replace.zeros(dx_M)
dx_F <- replace.zeros(dx_F)

# Fit model for US males using US females as benchmark
M <- model.OeppenC(data = dx_M,
                   data.B = dx_F,
                   x = x, y = y)
M

summary(M)
coef(M)
coef(M$benchmark)

# Plot observed and fitted values
plot(M, plotType = "observed")
plot(M, plotType = "fitted")

# Plot residuals
R <- residuals(M)
plot(R, plotType = "scatter")
plot(R, plotType = "colourmap")
plot(R, plotType = "signplot")

# Perform forecasts
P  <- predict(M, h = 16)
P

plot(P, plotType = "mean")
plot(P, plotType = "lower")
plot(P, plotType = "upper")
}
\seealso{
\code{\link{predict.Oeppen}}
\code{\link{plot.Oeppen}}
}
