% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RenshawHaberman.R
\name{model.RenshawHaberman}
\alias{model.RenshawHaberman}
\title{The Renshaw-Haberman Mortality Model imported from StMoMo package}
\usage{
model.RenshawHaberman(data, x, y, link = c("log", "logit"),
  cohortAgeFun = c("1", "NP"), approxConst = FALSE, radix = 1e+05,
  verbose = FALSE)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{link}{defines the link function and random component associated with 
the mortality model. \code{"log"} would assume that deaths follow a 
Poisson distribution and use a log link while \code{"logit"} would assume 
that deaths follow a Binomial distribution and a logit link.}

\item{cohortAgeFun}{defines the cohort age modulating parameter 
\eqn{\beta_x^{(0)}}. It can take values: \code{"NP"} for a non-parametric age
term or \code{"1"} for \eqn{\beta_x^{(0)}=1} (the default).}

\item{approxConst}{defines if the approximate identifiability constraint of 
Hunt and Villegas (2015) is applied or not. If \code{TRUE}, the output object 
is of class \code{rh} and subsequent model fitting is performed with 
\code{\link{fit.rh}}. If \code{FALSE}, the output object is of class 
\code{StMoMo} and subsequent model fitting is performed with 
\code{\link{fit.StMoMo}}.}

\item{radix}{Radix.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}
}
\description{
The Renshaw-Haberman mortality model is a Lee-Carter model with cohort effects.
}
\details{
\insertNoCite{renshaw2006}{MortalityForecast}
}
\examples{
# Data
x  <- 0:89
y  <- 2008:2014
mx <- HMD_male$mx$GBRTENW[paste(x), paste(y)]

# Fit the model
M <- model.RenshawHaberman(data = mx, x = x, y = y)
M
summary(M)

# Check residuals
R <- residuals(M)

plot(R, plotType = "scatter")
plot(R, plotType = "colourmap")
plot(R, plotType = "signplot")

# Forecast
P <- predict(M, h = 5)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{predict.RenshawHaberman}}
}
