% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_HyndmanUllah.R
\name{model_HyndmanUllah}
\alias{model_HyndmanUllah}
\title{The Functional Demographic Model}
\usage{
model_HyndmanUllah(data, x, y, order = 1, transform = TRUE, ...)
}
\arguments{
\item{data}{demogdata object. Output from read.demogdata.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{order}{Number of basis functions to fit.}

\item{transform}{If TRUE, the data are transformed with a Box-Cox transformation before the model is fitted.}

\item{...}{Extra arguments passed to \code{\link[ftsa]{ftsm}}.}
}
\value{
Object of class \dQuote{fdm} with the following components:
\item{label}{Name of country}
\item{age}{Ages from \code{data} object.}
\item{year}{Years from \code{data} object.}
\item{<series>}{Matrix of demographic data as contained in \code{data}. It takes the name given by the series argument.}
\item{fitted}{Matrix of fitted values.}
\item{residuals}{Residuals (difference between observed and fitted).}
\item{basis}{Matrix of basis functions evaluated at each age level (one column for each basis function). The first column is the fitted mean.}
\item{coeffs}{Matrix of coefficients (one column for each coefficient series). The first column are all ones.}
\item{mean.se}{Standard errors for the estimated mean function.}
\item{varprop}{Proportion of variation explained by each basis function.}
\item{weights}{Weight associated with each time period.}
\item{v}{Measure of variation for each time period.}
\item{type}{Data type (mortality, fertility, etc.)}
\item{y}{The data stored as a functional time series object.}
}
\description{
The Functional Demographic Model
}
\details{
\insertNoCite{hyndman2007}{MortalityForecast}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{model_LeeCarter}}
\code{\link{model_Oeppen}}
}
