% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LiLee.R
\name{model_LiLee}
\alias{model_LiLee}
\title{The Li-Lee Mortality Model}
\usage{
model_LiLee(data, benchmark, x = NULL, y = NULL, verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{benchmark}{A list containing multiple matrices with mortality data 
to be used as benchmark. Format: ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Fit the Li-Lee mortality model
}
\details{
\insertNoCite{li2005}{MortalityForecast}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{predict.LiLee}}
\code{\link{model_LeeCarter}}
}
