% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityForecast-package.r
\docType{package}
\name{MortalityAccuracy-package}
\alias{MortalityForecast}
\alias{MortalityForecast-package}
\title{MortalityForecast: Standard tools to compare and evaluate mortality forecasting methods}
\description{
Standard tools to compare and evaluate mortality forecasting methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpascariu/MortalityForecast}
  \item Report bugs at \url{https://github.com/mpascariu/MortalityForecast/issues}
}

}
\author{
\strong{Maintainer}: Marius D. Pascariu \email{rpascariu@outlook.com} [copyright holder]

}
