% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityForecast-data.R
\docType{data}
\name{MortalityForecast.data}
\alias{MortalityForecast.data}
\title{DATA in the package - for testing purposes}
\format{An object of class \code{list} of length 7.}
\source{
\href{https://www.mortality.org}{Human Mortality Database}
}
\usage{
MortalityForecast.data
}
\description{
Demographic data for England \& Wales male population between 1960 and 2016. 
The data is provided in the package for testing purposes only.
By the time you are using it, the data may be outdated. Download actual 
demographic data free of charge from \insertCite{hmd2018;textual}{MortalityForecast}. 
Once a username and a password is created on the 
\href{https://www.mortality.org}{website} the 
\href{https://CRAN.R-project.org/package=MortalityLaws}{MortalityLaws} 
R package can be used to extract data directly into your R console.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
