% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityForecast-package.R
\docType{package}
\name{MortalityForecast}
\alias{MortalityForecast-package}
\title{MortalityForecast: Standard Tools to Compare and Evaluate Various Mortality Forecasting Methods}
\description{
Standard tools to compare and evaluate mortality forecasting methods.
}
\details{
\insertNoCite{*}{MortalityForecast}
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpascariu/MortalityForecast}
  \item Report bugs at \url{https://github.com/mpascariu/MortalityForecast/issues}
}

}
\author{
\strong{Maintainer}: Marius D. Pascariu \email{rpascariu@outlook.com} [copyright holder]

Other contributors:
\itemize{
  \item Marie-Pier Bergeron-Boucher [contributor]
  \item Adam Lenart [contributor]
}

}
