% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_BBackTesting.R
\name{num_workers}
\alias{num_workers}
\title{Determine the number of cores to be used}
\usage{
num_workers(n.cores = NULL, res = NULL)
}
\arguments{
\item{n.cores}{Number of cores to be used in parallel computing;}

\item{res}{Restrict the number of cores. If there are only 2 operations to 
performe even if more cores are available 2 cores are used.}
}
\description{
Determine the number of cores to be used
}
\keyword{internal}
