% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{Oeppen.input.check}
\alias{Oeppen.input.check}
\title{Validate input values}
\usage{
Oeppen.input.check(X)
}
\arguments{
\item{X}{A list with input arguments provided in \code{\link{model_Oeppen}} function}
}
\description{
Validate input values
}
\keyword{internal}
