% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plot.doBackTesting}
\alias{plot.doBackTesting}
\title{Plot method for objects of the class doBackTesting}
\usage{
\method{plot}{doBackTesting}(x, filter.x = c(0, 60, 70, 80, 90, 100), ...)
}
\arguments{
\item{x}{An object of the class \code{doBackTesting}.}

\item{filter.x}{subset x vector}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Plot method for objects of the class doBackTesting
}
