% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODA_Plots.R
\name{plot.fitOeppen}
\alias{plot.fitOeppen}
\title{ggplot the observed and fitted values of a CoDa-LC mortality model}
\usage{
\method{plot}{fitOeppen}(x, plotType = c("fitted", "observed"), ny = 7,
  level = 80, ...)
}
\arguments{
\item{x}{An object of class \code{residMF}.}

\item{plotType}{The type of the plot. The alternatives are 
\code{"fitted"}, \code{"observed"}. Default: \code{"fitted"}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}

\item{level}{Sets the number of quantiles the data should be broken into. 
Default: 80.}

\item{...}{Other plotting parameters to be passed to the plotting 
functions. This can be used to control the appearance of the plots.}
}
\description{
ggplot the observed and fitted values of a CoDa-LC mortality model
}
\details{
Note: the output is a ggplot2 object. Therefore, one can add or 
modify the components of the figure.
}
\examples{
# For examples go to ?fitOeppen
}
