% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plot.getAccuracy}
\alias{plot.getAccuracy}
\title{Plot method for objects of the class \code{getAccuracy}}
\usage{
\method{plot}{getAccuracy}(x, fill = "#addd8e", lwd = 0.5, ...)
}
\arguments{
\item{x}{An object of the class \code{getAccuracy}.}

\item{fill}{Colour for filling the box of the best-performing method.}

\item{lwd}{Line width for marking the box of the best-performing method.}

\item{...}{Ignored.}
}
\description{
Plot method for objects of the class \code{getAccuracy}
}
\seealso{
\code{\link{getAccuracy}}
}
