% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_Plot_fit.R
\name{plot.MEM}
\alias{plot.MEM}
\title{ggplot the observed and fitted values of a Maximum-Entropy Mortality Model}
\usage{
\method{plot}{MEM}(x, plotType = c("fitted", "observed"), ny = 7,
  quant = c(0.1, 0.9), ...)
}
\arguments{
\item{x}{An object of class \code{residMF}.}

\item{plotType}{The type of the plot. The alternatives are 
\code{"fitted"}, \code{"observed"}. Default: \code{"fitted"}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}

\item{quant}{Sets the number of quantiles the data should be broken into. 
Default: c(0.1, 0.9).}

\item{...}{Other plotting parameters to be passed to the plotting 
functions. This can be used to control the appearance of the plots.}
}
\description{
ggplot the observed and fitted values of a Maximum-Entropy Mortality Model
}
\details{
Note: the output is a ggplot2 object. Therefore, one can add or 
modify the components of the figure.
}
\examples{
x  <- 0:110
y  <- 1965:2014
dx <- MortalityForecast.data$dx[paste(x), paste(y)]
M  <- MEM(dx, x, y, n = 5)

plot(M, plotType = "fitted")
plot(M, plotType = "observed")
}
