% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODA_Plots.R
\name{plot.predict.fitOeppen}
\alias{plot.predict.fitOeppen}
\title{ggplot the predicted values of a CoDa-LC mortality model}
\usage{
\method{plot}{predict.fitOeppen}(x, plotType = c("mean", "lower",
  "upper"), ny = 7, level = 80, ...)
}
\arguments{
\item{x}{An object of the class \code{\link{predict.fitOeppen}}.}

\item{plotType}{The type of the plot. The alternatives are 
\code{"mean", "lower", "upper"}. Default: \code{"mean"}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}

\item{level}{Sets the number of quantiles the data should be broken into. 
Default: 80.}

\item{...}{Other plotting parameters to be passed to the plotting 
functions. This can be used to control the appearance of the plots.}
}
\description{
ggplot the predicted values of a CoDa-LC mortality model
}
\examples{
# For examples go to ?predict.fitOeppen
}
