% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MEM.R
\name{plot.predict.MEM}
\alias{plot.predict.MEM}
\title{Plot the Predicted Age-at-Death Distribution}
\usage{
\method{plot}{predict.MEM}(x, y = NULL, plotType = c("mean", "lower",
  "upper", "raw_moments", "normalised_moments", "scaled_moments"),
  ny = 7, level = 80, ...)
}
\arguments{
\item{x}{An object of the class \code{\link{predict.MEM}}.}

\item{y}{An object of the class \code{\link{model.MEM}}. 
Needed only for plotting moments. See \code{plotType}.}

\item{plotType}{The type of the plot. The alternatives are 
\code{"mean", "lower", "upper", "raw_moments", "normalised_moments", 
"scaled_moments"}. Default: \code{"mean"}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}

\item{level}{Sets the number of quantiles the data should be broken into. 
Default: 80.}

\item{...}{Other plotting parameters to be passed to the plotting 
functions. This can be used to control the appearance of the plots.}
}
\description{
Plot the Predicted Age-at-Death Distribution
}
\examples{
# For examples go to ?model.MEM
}
\seealso{
\code{\link{model.MEM}}
}
