% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Oeppen.R
\name{plot.predict.Oeppen}
\alias{plot.predict.Oeppen}
\alias{plot.predict.OeppenC}
\title{Plot the Predicted Age-at-Death Distribution}
\usage{
\method{plot}{predict.Oeppen}(x, plotType = c("mean", "lower", "upper"),
  ny = 7, level = 80, ...)

\method{plot}{predict.OeppenC}(x, plotType = c("mean", "lower", "upper"),
  ny = 7, level = 80, ...)
}
\arguments{
\item{x}{An object generated using the \code{predict} function;}

\item{plotType}{The type of the plot. The alternatives are 
\code{"mean", "lower", "upper"}. Default: \code{"mean"}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}

\item{level}{Sets the number of quantiles the data should be broken into. 
Default: 80.}

\item{...}{Other plotting parameters to be passed to the plotting 
functions. This can be used to control the appearance of the plots.}
}
\description{
Plot the Predicted Age-at-Death Distribution
}
\examples{
# For examples go to ?model.Oeppen or ?model.OeppenC
}
\seealso{
\code{\link{model.Oeppen}}
\code{\link{model.OeppenC}}
}
