% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_predict.R
\name{predict.fitMaxEntMortality}
\alias{predict.fitMaxEntMortality}
\title{Predict Maximum-Entropy Mortality Model}
\usage{
\method{predict}{fitMaxEntMortality}(object, h, x.h = NULL, level = 95,
  jumpchoice = c("actual", "fit"), verbose = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{fitMaxEntMortality}}.}

\item{h}{Number of years to be forecast in the future.}

\item{x.h}{Numerical vector indicating the ages to be considered in 
reconstruction of the density over the forecast horizon. 
If \code{NULL}, the number of estimated data points is equal to the number 
of fitted data points. This argument can be used for example to estimate a 
density between 0 and 130 given the fact that the model was fitted on a 
dataset containing values for 0-100 only.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Predict Maximum-Entropy Mortality Model
}
\examples{
x  <- 0:110
y  <- 1985:2016
dx <- MortalityForecast.data$dx[paste(x), paste(y)]
M  <- fitMaxEntMortality(dx, x, y, n = 5)
P  <- predict(M, h = 16, x.h = 0:120)

plot(P, plotType = "mean")
plot(P, plotType = "lower")
plot(P, plotType = "upper")

plot(P, M, plotType = "raw_moments")
plot(P, M, plotType = "normalised_moments")
}
\seealso{
\code{\link{fitMaxEntMortality}}
}
