% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRW_predict.R
\name{predict.fitMRandomWalk}
\alias{predict.fitMRandomWalk}
\title{Forecast a Multivariate Random Walk Model}
\source{
The original implementation of this function was taken from StMoMo R package.
}
\usage{
\method{predict}{fitMRandomWalk}(object, h = 10, level = c(80, 95),
  ...)
}
\arguments{
\item{object}{An object of class \code{"fitMRandomWalk"}.}

\item{h}{Number of years to be forecast in the future.}

\item{level}{Significance level of the confidence interval.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
An object of the class \code{"predict.fitMRandomWalk"} with components:
\item{predicted.values}{ Data.frame with the central forecast.}
\item{conf.intervals}{ List with lower and upper limits for prediction intervals.}
\item{x}{ Numerical vector indicating the age classes in output.}
\item{y}{ Numerical vector indicating the years in output.}
}
\description{
Returns forecasts and other information for a Multivariate 
Random Walk Model.
}
\author{
Marius D. Pascariu
}
