% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MEM.R
\name{predict.MEM}
\alias{predict.MEM}
\title{Predict Maximum-Entropy Mortality Model}
\usage{
\method{predict}{MEM}(object, h, x.h = NULL, level = 95,
  jumpchoice = c("actual", "fit"), verbose = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{MEM};}

\item{h}{Number of years to be forecast in the future.}

\item{x.h}{Numerical vector indicating the ages to be considered in 
reconstruction of the density over the forecast horizon.
If \code{NULL}, the number of estimated data points is equal to the number 
of fitted data points. This argument can be used for example to estimate a 
density between 0 and 130 given the fact that the model was fitted on a 
dataset containing values for 0-100 only.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
The output is a list with the components:
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given 
 arguments;}
 \item{info}{Short details about the model;}
 \item{predicted.values}{A list containing the predicted d[x] values given 
 by the estimated model over the forecast horizon \code{h};}
 \item{conf.intervals}{Confidence intervals for the predcted values;}
 \item{predicted.raw.moments}{Predicted raw moments over the forecast horizon 
 \code{h};}
 \item{random.walk.model}{The estimated multivariate Random Walk model used 
 in the extrapolation of the moments;} 
 \item{x}{Vector of ages used in prediction;} 
 \item{y}{Vector of years used in prediction.}
}
\description{
Predict Maximum-Entropy Mortality Model
}
\examples{
x  <- 0:110
y  <- 1985:2016
dx <- HMD_male$dx$GBRTENW[paste(x), paste(y)]
M  <- model_MEM(dx, x, y, n = 5)
P  <- predict(M, h = 16, x.h = 0:120)

plot(P, plotType = "mean")
plot(P, plotType = "lower")
plot(P, plotType = "upper")

plot(P, M, plotType = "raw_moments")
plot(P, M, plotType = "normalised_moments")
}
\seealso{
\code{\link{model_MEM}}
}
