% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fitMortality.R
\name{predict.MortalityModels}
\alias{predict.MortalityModels}
\title{Forecast Mortality}
\usage{
\method{predict}{MortalityModels}(object, h, ci = 95,
  jumpchoice = c("actual", "fit"), ...)
}
\arguments{
\item{object}{An object of the class \code{MotalityModels} generated by the
\code{\link{getMortalityModels}} function.}

\item{h}{Number of years to be forecast in the future}

\item{ci}{Confidence level for prediction intervals.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{...}{Further argumets to be passed to other methods.}
}
\description{
Forecast Mortality
}
