% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRW_predict.R
\name{predict.MRW}
\alias{predict.MRW}
\title{Forecast a Multivariate Random Walk Model}
\usage{
\method{predict}{MRW}(object, h = 10, level = c(80, 95), ...)
}
\arguments{
\item{object}{An object of the class \code{"MWR"}.}

\item{h}{Number of periods for forecasting.}

\item{level}{Confidence level for prediction intervals.}

\item{...}{Ignored.}
}
\value{
An object of the class \code{"predict.MRW"} with components:
\item{predicted.values}{ Data.frame with the central forecast.}
\item{conf.intervals}{ List with lower and upper limits for prediction intervals.}
\item{x}{ Numerical vector indicating the age classes in output.}
\item{y}{ Numerical vector indicating the years in output.}
}
\description{
Returns forecasts and other information for a Multivariate 
Random Walk Model.
}
