% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{predict.Oeppen}
\alias{predict.Oeppen}
\title{Forecast the Age at Death Distribution using the Oeppen model.}
\usage{
\method{predict}{Oeppen}(object, h, order = NULL, include.drift = NULL,
  level = c(80, 95), jumpchoice = c("actual", "fit"), method = "ML",
  verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{Oeppen}.}

\item{h}{Number of years to be forecast in the future.}

\item{order}{A specification of the non-seasonal part of the ARIMA model: 
the three components (p, d, q) are the AR order, the degree of differencing, 
and the MA order. If \code{order = NULL}, the ARIMA order will be estimated 
automatically using the KPPS algorithm.}

\item{include.drift}{Logical. Should the ARIMA model include a linear drift term?
If \code{include.drift = NULL}, the model will be estimated automatically.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{method}{ARIMA fitting method: maximum likelihood or minimize conditional 
sum-of-squares. Options to use: conditional-sum-of-squares (\code{"CSS-ML"}), 
maximum likelihood (\code{"ML"}) and \code{"CSS"}.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Additional arguments to be passed to \code{\link[forecast]{Arima}}}
}
\value{
The output is an object of class \code{"predict.Oeppen"} with the components:
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given arguments;}
 \item{predicted.values}{A list containing the predicted values together
 with the associated prediction intervals given by the estimated 
 model over the forecast horizon \code{h};}
 \item{kt.arima}{An object of class \code{ARIMA} that contains all the
 components of the fitted time series model used in \code{kt} prediction;} 
 \item{kt}{The extrapolated \code{kt} parameters;}
 \item{conf.intervals}{Confidence intervals for the extrapolated \code{kt} 
 parameters;}
 \item{x}{Vector of ages used in prediction;} 
 \item{y}{Vector of years used in prediction;}
 \item{info}{Short details about the model.}
}
\description{
Forecast the Age at Death Distribution using the Oeppen model.
}
\examples{
# Example 1 ----------------------
x  <- 0:100
y  <- 1980:2014
dx <- HMD_male$dx$GBRTENW[paste(x), paste(y)]
M  <- model_Oeppen(dx, x, y)
P  <- predict(M, h = 16)

plot(P, plotType = "mean")
plot(P, plotType = "lower")
plot(P, plotType = "upper")

#' # Example 2 ----------------------
# One can specify manually the ARIMA order, a drift to be included or not 
# and the jump choice of the first forecast year.
P2 <- predict(M, h = 20, order = c(0,1,0), include.drift = TRUE, jumpchoice = "fit")

\dontrun{
# Example 3 ----------------------
# Compute life tables using forecast values using the MortalityLaws R package
library(MortalityLaws)
dx <- P$predicted.values
lt <- LifeTable(x = P$x, dx = dx)
}
}
\author{
Marius D. Pascariu, Marie-Pier Bergeron-Boucher and Jim Oeppen
}
