% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RenshawHaberman.R
\name{predict.RenshawHaberman}
\alias{predict.RenshawHaberman}
\title{Forecast age-specific death rates using the Renshaw-Haberman model}
\usage{
\method{predict}{RenshawHaberman}(object, h, order = c(2, 1, 0),
  include.drift = TRUE, level = c(80, 95), jumpchoice = c("actual",
  "fit"), verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{RenshawHaberman}.}

\item{h}{Number of years to be forecast in the future.}

\item{order}{A specification of the ARIMA model for the cohort effect: 
the three components (p, d, q) are the AR order, the degree of differencing, 
and the MA. Default: \code{c(1, 1, 0)}.}

\item{include.drift}{a logical value indicating if the ARIMA model should 
include a constant value. Default: \code{TRUE}.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Other arguments to be passed on to 
\code{\link[StMoMo]{forecast.fitStMoMo}}}
}
\description{
Forecast age-specific death rates using the Renshaw-Haberman model
}
\details{
\insertNoCite{renshaw2006}{MortalityForecast}
}
\examples{
# For examples go to ?model.RenshawHaberman
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{model.RenshawHaberman}}
}
