% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fitMortality.R
\name{predicted}
\alias{predicted}
\title{Get Predicted Values}
\usage{
predicted(object, type = c("qx", "mx", "dx", "lx", "Lx", "Tx", "ex"), ...)
}
\arguments{
\item{object}{An object of class \code{predict.MortalityModels}}

\item{type}{Specify the type of fitted values to be extracted.}

\item{...}{Further argumets to be passed to other methods.}
}
\description{
Get Predicted Values
}
