% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_fit.R
\name{print.coda}
\alias{print.coda}
\title{Print coda}
\usage{
\method{print}{coda}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"coda"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print coda
}
\keyword{internal}
