% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doForecasts.R
\name{print.doForecasts}
\alias{print.doForecasts}
\title{Print function for doForecasts}
\usage{
\method{print}{doForecasts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"doForecasts"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for doForecasts
}
\keyword{internal}
