% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_fit.R
\name{print.fitMaxEntMortality}
\alias{print.fitMaxEntMortality}
\title{Print function for Maximum-Entropy Mortality Model}
\usage{
\method{print}{fitMaxEntMortality}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{fitMaxEntMortality}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for Maximum-Entropy Mortality Model
}
\keyword{internal}
