% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODA_fit.R
\name{print.fitOeppen}
\alias{print.fitOeppen}
\title{Print fitOeppen}
\usage{
\method{print}{fitOeppen}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fitOeppen"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print fitOeppen
}
\keyword{internal}
