% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuracy.R
\name{print.getAccuracy}
\alias{print.getAccuracy}
\title{Print getAccuracy}
\usage{
\method{print}{getAccuracy}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{An object of the class \code{getAccuracy}}

\item{digits}{Number of digits to display.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Print getAccuracy
}
\keyword{internal}
