% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_HyndmanUllah.R
\name{print.HyndmanUllah}
\alias{print.HyndmanUllah}
\title{Print Hyndman-Ullah model}
\usage{
\method{print}{HyndmanUllah}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"HyndmanUllah"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Hyndman-Ullah model
}
\keyword{internal}
