% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{print.LeeCarter2}
\alias{print.LeeCarter2}
\title{Print Lee-Carter model}
\usage{
\method{print}{LeeCarter2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"LeeCarter2"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Lee-Carter model
}
\keyword{internal}
