% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LiLee.R
\name{print.LiLee}
\alias{print.LiLee}
\title{Print Li-Lee model}
\usage{
\method{print}{LiLee}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"LeeCarter2"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Li-Lee model
}
\keyword{internal}
