% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_fit.R
\name{print.MEM}
\alias{print.MEM}
\title{Print function for MEM method}
\usage{
\method{print}{MEM}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{MEM}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for MEM method
}
\keyword{internal}
