% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_MortalityForecasts.R
\name{print.MortalityForecasts}
\alias{print.MortalityForecasts}
\title{Print function for do.MortalityForecasts}
\usage{
\method{print}{MortalityForecasts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"MortalityForecasts"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for do.MortalityForecasts
}
\keyword{internal}
