% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_MortalityModels.R
\name{print.MortalityModels}
\alias{print.MortalityModels}
\title{Print function for do.MortalityModels}
\usage{
\method{print}{MortalityModels}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"MortalityModels"}}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Print function for do.MortalityModels
}
\keyword{internal}
