% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MRW.R
\name{print.MRW}
\alias{print.MRW}
\title{Print function for MRW method}
\usage{
\method{print}{MRW}(x, ...)
}
\arguments{
\item{x}{An object of class \code{MRW}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for MRW method
}
\keyword{internal}
