% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{print.Oeppen}
\alias{print.Oeppen}
\title{Print Oeppen}
\usage{
\method{print}{Oeppen}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"Oeppen"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Oeppen
}
\keyword{internal}
