% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_OeppenC.R
\name{print.OeppenC}
\alias{print.OeppenC}
\title{Print Coherent Oeppen}
\usage{
\method{print}{OeppenC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"OeppenC"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Coherent Oeppen
}
\keyword{internal}
