% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_predict.R
\name{print.predict.coda}
\alias{print.predict.coda}
\title{Print predict.coda}
\usage{
\method{print}{predict.coda}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.coda"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print predict.coda
}
\keyword{internal}
