% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_predict.R
\name{print.predict.fitMaxEntMortality}
\alias{print.predict.fitMaxEntMortality}
\title{Print function for \code{\link{predict.fitMaxEntMortality}}}
\usage{
\method{print}{predict.fitMaxEntMortality}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{fitMaxEntMortality}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for \code{\link{predict.fitMaxEntMortality}}
}
\keyword{internal}
