% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODA_predict.R
\name{print.predict.fitOeppen}
\alias{print.predict.fitOeppen}
\title{Print predict.fitOeppen}
\usage{
\method{print}{predict.fitOeppen}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.fitOeppen"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print predict.fitOeppen
}
\keyword{internal}
