% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_HyndmanUllah.R
\name{print.predict.HyndmanUllah}
\alias{print.predict.HyndmanUllah}
\title{Print function}
\usage{
\method{print}{predict.HyndmanUllah}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.HyndmanUllah"};}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function
}
\keyword{internal}
