% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{print.predict.LeeCarter}
\alias{print.predict.LeeCarter}
\title{Print function}
\usage{
\method{print}{predict.LeeCarter}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.LeeCarter"};}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function
}
\keyword{internal}
