% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter2.R
\name{print.predict.LeeCarter2}
\alias{print.predict.LeeCarter2}
\title{Print function}
\usage{
\method{print}{predict.LeeCarter2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.LeeCarter2"};}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function
}
\keyword{internal}
