% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LiLee.R
\name{print.predict.LiLee}
\alias{print.predict.LiLee}
\title{Print function}
\usage{
\method{print}{predict.LiLee}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.LiLee"};}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function
}
\keyword{internal}
