% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MRW.R
\name{print.predict.MRW}
\alias{print.predict.MRW}
\title{Print function}
\usage{
\method{print}{predict.MRW}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.MRW"};}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function
}
\keyword{internal}
