% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{print.predict.Oeppen}
\alias{print.predict.Oeppen}
\title{Print predict.Oeppen}
\usage{
\method{print}{predict.Oeppen}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.Oeppen"};}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print predict.Oeppen
}
\keyword{internal}
