% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_OeppenC.R
\name{print.predict.OeppenC}
\alias{print.predict.OeppenC}
\title{Print predict.OeppenC}
\usage{
\method{print}{predict.OeppenC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.OeppenC"};}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print predict.OeppenC
}
\keyword{internal}
