% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_fit.R
\name{print.summary.coda}
\alias{print.summary.coda}
\title{Print summary.coda}
\usage{
\method{print}{summary.coda}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.coda"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary.coda
}
\keyword{internal}
