% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODA_fit.R
\name{print.summary.fitOeppen}
\alias{print.summary.fitOeppen}
\title{Print summary.fitOeppen}
\usage{
\method{print}{summary.fitOeppen}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.fitOeppen"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary.fitOeppen
}
\keyword{internal}
