% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResiduals.R
\name{print.summary.getResiduals}
\alias{print.summary.getResiduals}
\title{Print summary.getResiduals}
\usage{
\method{print}{summary.getResiduals}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.getResiduals}}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Print summary.getResiduals
}
\keyword{internal}
