% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{print.summary.LeeCarter}
\alias{print.summary.LeeCarter}
\title{Print summary.LeeCarter}
\usage{
\method{print}{summary.LeeCarter}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.LeeCarter"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary.LeeCarter
}
\keyword{internal}
