% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{print.summary.LeeCarter2}
\alias{print.summary.LeeCarter2}
\title{Print summary.LeeCarter2}
\usage{
\method{print}{summary.LeeCarter2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.LeeCarter2"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary.LeeCarter2
}
\keyword{internal}
