% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LiLee.R
\name{print.summary.LiLee}
\alias{print.summary.LiLee}
\title{Print summary.LiLee}
\usage{
\method{print}{summary.LiLee}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.LiLee"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary.LiLee
}
\keyword{internal}
