% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{print.summary.Oeppen}
\alias{print.summary.Oeppen}
\title{Print summary.Oeppen}
\usage{
\method{print}{summary.Oeppen}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.Oeppen"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary.Oeppen
}
\keyword{internal}
