% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_OeppenC.R
\name{print.summary.OeppenC}
\alias{print.summary.OeppenC}
\title{Print summary.OeppenC}
\usage{
\method{print}{summary.OeppenC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.OeppenC"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary.OeppenC
}
\keyword{internal}
