% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_HyndmanUllah.R, R/model_LeeCarter.R,
%   R/model_LiLee.R, R/model_MEM.R, R/model_MRW.R, R/model_Oeppen.R,
%   R/model_OeppenC.R, R/model_RenshawHaberman.R, R/utils.R
\name{print.HyndmanUllah}
\alias{print.HyndmanUllah}
\alias{print.predict.HyndmanUllah}
\alias{print.LeeCarter}
\alias{print.summary.LeeCarter}
\alias{print.predict.LeeCarter}
\alias{print.LiLee}
\alias{print.summary.LiLee}
\alias{print.predict.LiLee}
\alias{print.MEM}
\alias{print.predict.MEM}
\alias{print.MRW}
\alias{print.predict.MRW}
\alias{print.Oeppen}
\alias{print.summary.Oeppen}
\alias{print.predict.Oeppen}
\alias{print.OeppenC}
\alias{print.summary.OeppenC}
\alias{print.predict.OeppenC}
\alias{print.RenshawHaberman}
\alias{print.summary.RenshawHaberman}
\alias{print.predict.RenshawHaberman}
\alias{print_default}
\alias{print_predict_default}
\title{Print function for mortality objects}
\usage{
\method{print}{HyndmanUllah}(x, ...)

\method{print}{predict.HyndmanUllah}(x, ...)

\method{print}{LeeCarter}(x, ...)

\method{print}{summary.LeeCarter}(x, ...)

\method{print}{predict.LeeCarter}(x, ...)

\method{print}{LiLee}(x, ...)

\method{print}{summary.LiLee}(x, ...)

\method{print}{predict.LiLee}(x, ...)

\method{print}{MEM}(x, ...)

\method{print}{predict.MEM}(x, ...)

\method{print}{MRW}(x, ...)

\method{print}{predict.MRW}(x, ...)

\method{print}{Oeppen}(x, ...)

\method{print}{summary.Oeppen}(x, ...)

\method{print}{predict.Oeppen}(x, ...)

\method{print}{OeppenC}(x, ...)

\method{print}{summary.OeppenC}(x, ...)

\method{print}{predict.OeppenC}(x, ...)

\method{print}{RenshawHaberman}(x, ...)

\method{print}{summary.RenshawHaberman}(x, ...)

\method{print}{predict.RenshawHaberman}(x, ...)

print_default(x, ...)

print_predict_default(x, ...)
}
\arguments{
\item{x}{A mortality object;}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for mortality objects
}
\keyword{internal}
