% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_predict_default}
\alias{print_predict_default}
\title{Default print function for predict methods}
\usage{
print_predict_default(x, ...)
}
\arguments{
\item{x}{An predict object;}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Default print function for predict methods
}
\keyword{internal}
