% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_value_zero}
\alias{replace_value_zero}
\title{If there are death rates are equal to zero they will be replaced
using the multiplicative replacement strategy.}
\usage{
replace_value_zero(mx, radix = 1e+05)
}
\arguments{
\item{mx}{Matrix of death rates}

\item{radix}{radix. Default: 1e5.}
}
\description{
If there are death rates are equal to zero they will be replaced
using the multiplicative replacement strategy.
}
\examples{
# x  <- 0:25
# y  <- 2005:2016
# mx <- HMD_male$mx$DNK[paste(x), paste(y)]
# mx == 0
# 
# new.mx <- replace_value_zero(mx)
# new.mx == 0
# 
# sum(mx) == sum(new.mx)
# (mx - new.mx)/mx * 100
}
\keyword{internal}
