% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_fit.R
\name{residuals.coda}
\alias{residuals.coda}
\title{Residuals of the CoDa Mortality Model}
\usage{
\method{residuals}{coda}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"coda"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Residuals of the CoDa Mortality Model
}
\keyword{internal}
