% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRW_fit.R
\name{residuals.fitMRandomWalk}
\alias{residuals.fitMRandomWalk}
\title{Residuals of a Multivariate Random-Walk Model}
\usage{
\method{residuals}{fitMRandomWalk}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{fitMRandomWalk}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Computed deviance residuals for a Multivariate Random-Walk model .
}
\seealso{
\code{\link{fitMRandomWalk}}
}
