% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODA_fit.R
\name{residuals.fitOeppen}
\alias{residuals.fitOeppen}
\title{Residuals of the CoDa-LC Mortality Model}
\usage{
\method{residuals}{fitOeppen}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fitOeppen"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Residuals of the CoDa-LC Mortality Model
}
