% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_HyndmanUllah.R
\name{residuals.HyndmanUllah}
\alias{residuals.HyndmanUllah}
\title{Residuals HyndmanUllah}
\usage{
\method{residuals}{HyndmanUllah}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"HyndmanUllah"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Residuals HyndmanUllah
}
