% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{residuals.LeeCarter}
\alias{residuals.LeeCarter}
\title{Residuals of the Lee-Carter Mortality Model}
\usage{
\method{residuals}{LeeCarter}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"LeeCarter"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Residuals of the Lee-Carter Mortality Model
}
