% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{residuals.LeeCarter2}
\alias{residuals.LeeCarter2}
\title{Residuals of the Lee-Carter Mortality Model}
\usage{
\method{residuals}{LeeCarter2}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"LeeCarter2"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Residuals of the Lee-Carter Mortality Model
}
