% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LiLee.R
\name{residuals.LiLee}
\alias{residuals.LiLee}
\title{Residuals of the Li-Lee Mortality Model}
\usage{
\method{residuals}{LiLee}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"LeeCarter2"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Residuals of the Li-Lee Mortality Model
}
