% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_fit.R
\name{residuals.MEM}
\alias{residuals.MEM}
\title{Residuals of the Maximum-Entropy Mortality Model}
\usage{
\method{residuals}{MEM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{MEM}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Computed deviance residuals for a fitted Maximum-Entropy Mortality (MEM) model.
}
\examples{
x  <- 0:110
y  <- 1965:2014
dx <- MortalityForecast.data$dx[paste(x), paste(y)]
M  <- MEM(dx, x, y, n = 5)
residuals(M)
}
