% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MRW.R
\name{residuals.MRW}
\alias{residuals.MRW}
\title{Residuals of a Multivariate Random-Walk Model}
\usage{
\method{residuals}{MRW}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MRW}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Computed deviance residuals for a Multivariate Random-Walk model .
}
\seealso{
\code{\link{fit_MRW}}
}
