% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_HyndmanUllah.R, R/model_LeeCarter.R,
%   R/model_LiLee.R, R/model_MEM.R, R/model_MRW.R, R/model_Oeppen.R,
%   R/model_OeppenC.R
\name{residuals.HyndmanUllah}
\alias{residuals.HyndmanUllah}
\alias{residuals.LeeCarter}
\alias{residuals.LiLee}
\alias{residuals.MEM}
\alias{residuals.MRW}
\alias{residuals.Oeppen}
\alias{residuals.OeppenC}
\title{Extract Model Residuals}
\usage{
\method{residuals}{HyndmanUllah}(object, ...)

\method{residuals}{LeeCarter}(object, ...)

\method{residuals}{LiLee}(object, ...)

\method{residuals}{MEM}(object, ...)

\method{residuals}{MRW}(object, ...)

\method{residuals}{Oeppen}(object, ...)

\method{residuals}{OeppenC}(object, ...)
}
\arguments{
\item{object}{A fitted mortality model}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Extract Model Residuals
}
\examples{
# See examples in the main functions linked above.
}
\seealso{
\code{\link{model.HyndmanUllah}}
\code{\link{model.LeeCarter}}
\code{\link{model.LiLee}}
\code{\link{model.MEM}}
\code{\link{model.MRW}}
\code{\link{model.Oeppen}}
\code{\link{model.OeppenC}}
}
