% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_OeppenC.R
\name{residuals.OeppenC}
\alias{residuals.OeppenC}
\title{Residuals of the Coherent Oeppen Mortality Model}
\usage{
\method{residuals}{OeppenC}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"OeppenC"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Residuals of the Coherent Oeppen Mortality Model
}
