% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{scatterplotAPC}
\alias{scatterplotAPC}
\title{Do a scatter plot of a matrix according to age-period-cohorts}
\usage{
scatterplotAPC(mat, ages, years, plotAge = TRUE, plotYear = TRUE,
  plotCohort = TRUE, zeroLine = TRUE, ...)
}
\arguments{
\item{mat}{Matrix with the data to plot.}

\item{ages}{Ages corresponding to the rows in \code{mat}.}

\item{years}{Years corresponding to the columns in \code{mat}.}

\item{plotAge}{Logical value indicating if the age scatter plot should be 
produced.}

\item{plotYear}{Logical value indicating if the calendar year scatter plot 
should be produced.}

\item{plotCohort}{Logical value indicating if the cohort scatter plot 
should be produced.}

\item{zeroLine}{Logical value indicating if a horizontal line at zero
should be plotted.}

\item{...}{Other plotting parameters to be passed to the plotting 
functions. This can be used to control the appearance of the plots.}
}
\description{
Do a scatter plot of a matrix according to age-period-cohorts
}
\keyword{internal}
