% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_fit.R
\name{summary.coda}
\alias{summary.coda}
\title{Summary coda}
\usage{
\method{summary}{coda}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"coda"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary coda
}
\keyword{internal}
