% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODA_fit.R
\name{summary.fitOeppen}
\alias{summary.fitOeppen}
\title{Summary fitOeppen}
\usage{
\method{summary}{fitOeppen}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fitOeppen"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary fitOeppen
}
\keyword{internal}
