% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Fitted.R
\name{summary.get.Fitted}
\alias{summary.get.Fitted}
\title{Summary for get.Fitted}
\usage{
\method{summary}{get.Fitted}(object, ..., digits = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{get.Fitted}}.}

\item{...}{Arguments to be passed to or from other methods.}

\item{digits}{Number of digits to display.}
}
\description{
Summary for get.Fitted
}
\keyword{internal}
