% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitted.R
\name{summary.getFitted}
\alias{summary.getFitted}
\title{Summary for getFitted}
\usage{
\method{summary}{getFitted}(object, ..., digits = max(4L,
  getOption("digits") - 2L))
}
\arguments{
\item{object}{An object of class \code{\link{getFitted}}.}

\item{...}{Arguments to be passed to or from other methods.}

\item{digits}{Number of digits to display.}
}
\description{
Summary for getFitted
}
\keyword{internal}
