% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{summary.LeeCarter}
\alias{summary.LeeCarter}
\title{Summary LeeCarter}
\usage{
\method{summary}{LeeCarter}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"LeeCarter"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary LeeCarter
}
\keyword{internal}
