% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{summary.LeeCarter2}
\alias{summary.LeeCarter2}
\title{Summary LeeCarter2}
\usage{
\method{summary}{LeeCarter2}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"LeeCarter2"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary LeeCarter2
}
\keyword{internal}
