% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LiLee.R
\name{summary.LiLee}
\alias{summary.LiLee}
\title{Summary LiLee}
\usage{
\method{summary}{LiLee}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"LeeCarter2"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary LiLee
}
\keyword{internal}
