% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{summary.Oeppen}
\alias{summary.Oeppen}
\title{Summary Oeppen}
\usage{
\method{summary}{Oeppen}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"Oeppen"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary Oeppen
}
\keyword{internal}
