% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_OeppenC.R
\name{summary.OeppenC}
\alias{summary.OeppenC}
\title{Summary Coherent Oeppen}
\usage{
\method{summary}{OeppenC}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"OeppenC"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary Coherent Oeppen
}
\keyword{internal}
