% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whatsYourName}
\alias{whatsYourName}
\title{Return the name of the object as.character}
\usage{
whatsYourName(x)
}
\arguments{
\item{x}{Any object.}
}
\description{
Return the name of the object as.character
}
\keyword{internal}
