% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_MortalityModels.R
\name{which_model_is_in}
\alias{which_model_is_in}
\title{Determine which model is included in input}
\usage{
which_model_is_in(X, M = c("MRW", "MRWD", "LeeCarter", "HyndmanUllah",
  "RenshawHaberman", "Oeppen", "MEM2", "MEM3", "MEM4", "MEM5", "MEM6",
  "MEM7", "LiLee", "OeppenC"))
}
\arguments{
\item{X}{Models in input}

\item{M}{All models}
}
\description{
Determine which model is included in input
}
\keyword{internal}
