% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wide.list.2.long.df}
\alias{wide.list.2.long.df}
\title{List containing wide tables to 1 long table}
\usage{
wide.list.2.long.df(data, x, filter.x = NULL, ...)
}
\arguments{
\item{data}{Wide table}

\item{x}{vector of ages}

\item{filter.x}{subset x vector}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
List containing wide tables to 1 long table
}
\keyword{internal}
