% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wide2long}
\alias{wide2long}
\title{Wide table to long table}
\usage{
wide2long(data, x = NULL, filter.x = NULL, ...)
}
\arguments{
\item{data}{Wide table}

\item{x}{vector of ages}

\item{filter.x}{subset x vector}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Wide table to long table
}
\keyword{internal}
