% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBBackTesting.R
\name{buildScenarios}
\alias{buildScenarios}
\title{Build Scenarios for doBackTESTING}
\usage{
buildScenarios(y, strategy = c(f = 20, h = 20, s = 2))
}
\arguments{
\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{strategy}{Fitting-Forecasting strategy. Format: numerical vector.
The strategy \code{c(f, h, s)} consists in \code{f} number of years to use in 
fitting, \code{h} number of years to forecast and step \code{s} to change the
fitting-forecasting window.}
}
\description{
Build Scenarios for doBackTESTING
}
\keyword{internal}
