% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalAccuracy.R
\name{computeAccuracy}
\alias{computeAccuracy}
\title{Get Measures of Forecast Accuracy}
\usage{
computeAccuracy(u, u.hat, b, measures = NULL, xa = NULL, ya = NULL,
  na.rm = TRUE)
}
\arguments{
\item{u}{Validation dataset.}

\item{u.hat}{Out-sample forecast data.}

\item{b}{Benchmark forecast data. Usualy a naive or Random-Walk w drift forecast.}

\item{measures}{What accurracy measure to compute? Various alternatives are 
available, \itemize{
 \item{Mean error measures: } \code{"ME", "MAE", "MAPE", "sMAPE", "sMRAE", "MASE"};
 \item{Median error measures: } \code{"MdE", "MdAE", "MdAPE", "sMdAPE", "sMdRAE", "MdASE"};
 \item{Squared error measures: } \code{"MSE", "RMSE", "RMSPE", "RMdSPE"};
 \item{Geometric mean measure for positive errors: } \code{"GMRAE"}.}
 If \code{measures = NULL} all the measures will be computed.}

\item{xa}{Ages to be considered in model accuracy evaluation. It can be used 
to calculate the measures on a subset of the results. If \code{xa = NULL} 
(default) the entire age-range in input is considered.}

\item{ya}{Years to be considered in accuracy computation. Default: \code{ya = NULL}.}

\item{na.rm}{A logical value indicating whether NA values should be stripped 
before the computation proceeds.}
}
\description{
Get Measures of Forecast Accuracy
}
\details{
See \insertCite{hyndman2006;textual}{MortalityForecast} for a 
comprehensive discussion of the accuracy measures.
}
\references{
\insertAllCited{}
}
\author{
Marius D. Pascariu
}
\keyword{internal}
