% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBackTesting.R
\name{computeAccuracy}
\alias{computeAccuracy}
\title{Get Measures of Forecast Accuracy}
\source{
Hyndman and Koehler, 2006
}
\usage{
computeAccuracy(u, u.hat, b, na.rm = TRUE)
}
\arguments{
\item{u}{Observed data}

\item{u.hat}{In-sample Forecast data}

\item{b}{Benchmark forecast data. Usualy a naive or rwd forecast.}

\item{na.rm}{A logical value indicating whether NA values should be stripped 
before the computation proceeds.}

\item{what}{What accurracy measure to compute? If \code{what = NULL} all 
15 measures will be generated.}
}
\description{
Get Measures of Forecast Accuracy
}
\keyword{internal}
