% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuracy.R
\name{computeAccuracy}
\alias{computeAccuracy}
\title{Get Measures of Forecast Accuracy}
\source{
Hyndman and Koehler, 2006
}
\usage{
computeAccuracy(u, u.hat, b, xa = NULL, ya = NULL, measures,
  na.rm = TRUE)
}
\arguments{
\item{u}{Validation dataset.}

\item{u.hat}{In-sample forecast data.}

\item{b}{Benchmark forecast data. Usualy a naive or Random-Walk forecast.}

\item{xa}{Ages to be considered in accuracy computation. It can be used to 
calculate the measures on a subset of the results. If \code{x = NULL} 
(default) the entire age-range in \code{u} is considered.}

\item{ya}{Years to be considered in accuracy computation. Default: \code{NULL}.}

\item{measures}{What accurracy measure to compute? Various alternatives are 
available, \itemize{
\item{Mean error measures: } \code{"ME", "MAE", "MAPE", "sMAPE", "MRAE", "MASE"};
\item{Median error measures: } \code{"MdE", "MdAE", "MdAPE", "sMdAPE", "MdRAE", "MdASE"};
\item{Squared error measures: } \code{"MSE", "RMSE", "RMSPE", "RMdSPE"};
\item{Geometric mean measure for positive errors: } \code{"GMRAE"}.}}

\item{na.rm}{A logical value indicating whether NA values should be stripped 
before the computation proceeds.}
}
\description{
Get Measures of Forecast Accuracy
}
\keyword{internal}
