% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_moments.R
\name{convertRM}
\alias{convertRM}
\title{Convert Raw Moments to Central or Normalized moments}
\usage{
convertRM(data, to = c("central", "normalized"))
}
\arguments{
\item{data}{Data.frame with raw moments.}

\item{to}{What type of statistical moments do you want to obtain? 
Three types of moments can be obtained: \code{"raw", "central", "normalized"}.}
}
\description{
Convert Raw Moments to Central or Normalized moments
}
\keyword{internal}
