% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MEM.R
\name{correct_jump_off}
\alias{correct_jump_off}
\title{Jump-off correction}
\usage{
correct_jump_off(X, object, jumpchoice, h)
}
\arguments{
\item{X}{matrix of fitted values}

\item{object}{An object of class \code{\link{fit_MEM}}.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{h}{Number of years to be forecast in the future.}
}
\description{
Jump-off correction
}
\keyword{internal}
