% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_BackTesting.R
\name{do.BackTesting}
\alias{do.BackTesting}
\title{Perform In-Sample Testing of Mortality Forecasts Over One Time Period}
\usage{
do.BackTesting(data, x, y.fit, y.for, data.in = c("qx", "mx", "dx",
  "lx"), models, level = 95, jumpchoice = c("actual", "fit"),
  verbose = FALSE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column. This object should
contain the data to be used in fitting (training) and validation process 
as well.}

\item{x}{Numerical vector indicating the ages in input \code{data}. 
Optional. Default: \code{NULL}.}

\item{y.fit}{Years to be considered in fitting.}

\item{y.for}{Years to be forecast.}

\item{data.in}{Specify the type of input \code{data}. Various life table 
indices are accepted: \code{"qx", "mx", "dx", "lx"}.}

\item{models}{One or several mortality models to be estimated. 
The following options are available: \itemize{
  \item{\code{"MRW"}} -- The Multivariate Random-Walk (w/o Drift);
  \item{\code{"MRWD"}} -- The Multivariate Random-Walk with Drift;
  \item{\code{"LeeCarter"}} -- The Lee-Carter Mortality Model;
  \item{\code{"LiLee"}} -- The Li-Lee Mortality Model;
  \item{\code{"HyndmanUllah"}} -- The Hyndman-Ullah Mortality Model;
  \item{\code{"Oeppen"}} -- The Oeppen Mortality Model;
  \item{\code{"OeppenC"}} -- The Coherent Oeppen Mortality Model;
  \item{\code{"MEM2"}} -- The Maximum Entropy Mortality Model of order 2;
  \item{\code{"MEM3"}} -- The Maximum Entropy Mortality Model of order 3;
  \item{\code{"MEM4"}} -- The Maximum Entropy Mortality Model of order 4;
  \item{\code{"MEM5"}} -- The Maximum Entropy Mortality Model of order 5;
  \item{\code{"MEM6"}} -- The Maximum Entropy Mortality Model of order 6;
  \item{\code{"MEM7"}} -- The Maximum Entropy Mortality Model of order 7;
}}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Perform In-Sample Testing of Mortality Forecasts Over One Time Period
}
\examples{

x = 0:98              # Ages
y1 = 1980:2000        # Training period
y2 = 2001:2016        # Validation period
y  = c(y1, y2)
h = max(y2) - max(y1) # Forecasting horizon

D <- HMD_male$dx$GBRTENW[paste(x), paste(y)] # DATA

# Select various mortality models
MM <- c("MRWD", "Oeppen", "MEM6")
# Fit & Forecast the models 
B <- do.BackTesting(data = D, x = x,
                    y.fit = y1, y.for = y2,
                    data.in = "dx",
                    models = MM)

# Compute accuracy measures
# The measures can be computed for different indicators. Even if it is not 
# impossible to get a different classification and ranking the 
# outcome should be in general the same.
evalAccuracy(B, data.out = "mx")
evalAccuracy(B, data.out = "qx")
evalAccuracy(B, data.out = "dx")

# Rank the model's performance.
A <- evalAccuracy(B, data.out = "ex")
A
R <- do.Ranking(A)
R

# Visualize the results
plot(B, data.out = "mx", facet = "x")
plot(B, data.out = "mx", facet = "y") 
}
\seealso{
\code{\link{do.BBackTesting}}
\code{\link{evalAccuracy.BackTesting}}
}
