% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBackTesting.R
\name{doBackTesting}
\alias{doBackTesting}
\title{Performe back-testing}
\usage{
doBackTesting(Tdata, object, data.type = c("qx", "mx", "dx"), type = c("qx",
  "mx", "dx", "lx", "Lx", "Tx", "ex"), ...)
}
\arguments{
\item{Tdata}{Test data}

\item{object}{An object of the class \code{doForecasts}.}

\item{data.type}{Type of data in 'data' argument. Options: 
\code{"qx", "mx", "dx"}.}

\item{type}{Specify the type of values to be extracted.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Performe back-testing
}
