% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBackTesting.R
\name{doBackTesting}
\alias{doBackTesting}
\title{Perform In-Sample Testing of Mortality Forecasts Over One Time Period}
\usage{
doBackTesting(data, x, y.fit, y.for, data.in = c("qx", "mx", "dx", "lx"),
  data.out = c("qx", "mx", "dx", "lx", "Lx", "Tx", "ex"),
  models = c("MRWD", "LC"), measures = c("ME", "MAE", "MAPE", "sMAPE",
  "MRAE", "MASE"), xa = x[-length(x)], ya = NULL, level = 95,
  jumpchoice = c("actual", "fit"), verbose = FALSE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column. This object should
contain the data to be used in fitting (traning) and validation process as well.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y.fit}{Years to be considered in fitting.}

\item{y.for}{Years to be forecast.}

\item{data.in}{Specify the type of input \code{data}. Various life table 
indices are accepted: \code{"qx", "mx", "dx", "lx"}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{models}{Mortality models to be evaluated.}

\item{measures}{What accurracy measure to compute? Various alternatives are 
available, \itemize{
 \item{Mean error measures: } \code{"ME", "MAE", "MAPE", "sMAPE", "MRAE", "MASE"};
 \item{Median error measures: } \code{"MdE", "MdAE", "MdAPE", "sMdAPE", "MdRAE", "MdASE"};
 \item{Squared error measures: } \code{"MSE", "RMSE", "RMSPE", "RMdSPE"};
 \item{Geometric mean measure for positive errors: } \code{"GMRAE"}.}}

\item{xa}{Ages to be considered in model accuracy evaluation. It can be used 
to calculate the measures on a subset of the results. If \code{xa = NULL} 
the entire age-range in \code{x} is considered. Default: \code{xa = x[-length(x)]}.
In default mode the last row of results (last row in the life table) is 
excluded from evaluation. Because of different methods applied to force closing of 
the life table (input-specific), the errors can be "false positives" in the last row.}

\item{ya}{Years to be considered in accuracy computation. If \code{ya = NULL} 
(default) the entire period in \code{y} is considered.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Perform In-Sample Testing of Mortality Forecasts Over One Time Period
}
\examples{
x = 0:98              # Ages
y1 = 1980:2000        # Training period
y2 = 2001:2016        # Validation period
y  = c(y1, y2)
h = max(y2) - max(y1) # Forecasting horizon

D <- MortalityForecast.data$dx[paste(x), paste(y)] # DATA

MM <- c("MRWD", "LC", "CoDa", "MEM6")
B.ex <- doBackTesting(data = D, x = x, 
                      y.fit = y1, y.for = y2, 
                      data.in = "dx", 
                      data.out = "ex", 
                      models = MM)
B.ex$accuracy

plot(B.ex, facet = "x")
plot(B.ex, facet = "y")  
}
\seealso{
\code{\link{doBBackTesting}}
\code{\link{doMortalityModels}}
\code{\link{doForecasts}}
\code{\link{getAccuracy}}
}
