% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBackTesting.R
\name{doBackTesting}
\alias{doBackTesting}
\title{Perform in-sample testing of mortality forecasts}
\usage{
doBackTesting(data, x, y.fit, y.for, data.type, what, level = 95,
  jumpchoice = "actual", xa = NULL, ya = NULL, measures = c("ME",
  "MAE", "MAPE", "sMAPE", "MRAE", "MASE"), models = c("LC", "FDM",
  "PLAT", "CoDa", "M4", "M4X", "M5", "M5X", "M6", "M6X"), ...)
}
\arguments{
\item{data}{Mortality data for fitting of the models and for validation; 
that is the data coresponding to the horizon to be predicted.}

\item{x}{Vector of ages.}

\item{y.fit}{Years to be considered in fitting.}

\item{y.for}{Years to be forecast.}

\item{data.type}{Type of data in 'data' argument. Options: 
\code{"qx", "mx", "dx", "lx"}.}

\item{what}{Specify the type of values to be extracted.}

\item{level}{Confidence level for prediction intervals.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{xa}{Ages to be considered in accuracy computation. It can be used to 
calculate the measures on a subset of the results. If \code{x = NULL} 
(default) the entire age-range in \code{u} is considered.}

\item{ya}{Years to be considered in accuracy computation. Default: \code{NULL}.}

\item{measures}{What accurracy measure to compute? Various alternatives are 
available, \itemize{
\item{Mean error measures: } \code{"ME", "MAE", "MAPE", "sMAPE", "MRAE", "MASE"};
\item{Median error measures: } \code{"MdE", "MdAE", "MdAPE", "sMdAPE", "MdRAE", "MdASE"};
\item{Squared error measures: } \code{"MSE", "RMSE", "RMSPE", "RMdSPE"};
\item{Geometric mean measure for positive errors: } \code{"GMRAE"}.}}

\item{models}{Mortality models to be evaluated.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Perform in-sample testing of mortality forecasts
}
\examples{
x = 0:98              # Ages
y1 = 1980:2000        # Training period
y2 = 2001:2016        # Validation period
y  = c(y1, y2)
h = max(y2) - max(y1) # Forecasting horizon

D <- MortalityForecast.data$dx[paste(x), paste(y)] # DATA

MM <- c("MRW", "MRWD", "LC", "CoDa", "M6")
B.ex <- doBackTesting(data = D, x = x, 
                      y.fit = y1, y.for = y2, 
                      data.type = "dx", 
                      what = "ex", 
                      models = MM)
B.ex$accuracy

plot(B.ex, facet = "x")
plot(B.ex, facet = "y")  
}
