% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doForecasts.R
\name{doForecasts}
\alias{doForecasts}
\title{Perform Multiple Forecast of Mortality Using Various Models}
\usage{
doForecasts(object, h, level = 95, jumpchoice = c("actual", "fit"),
  verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{MotalityModels} generated by the
\code{\link{doMortalityModels}} function.}

\item{h}{Number of years to be forecast in the future.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Perform Multiple Forecast of Mortality Using Various Models
}
\examples{
x  <- 0:100
y  <- 1985:1999
h  <- 17
D  <- MortalityForecast.data$dx[paste(x), paste(y)]
MM <- c("MRWD", "LeeCarter", "HyndmanUllah", "CoDa", "MEM6")

M <- doMortalityModels(data = D, x, y, data.in = "dx", models = MM)
P <- doForecasts(M, h, level = 95, jumpchoice = "actual")

pex <- getForecasts(P, data.out = "ex")
}
