% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMortality.R
\name{doMortalityModels}
\alias{doMortalityModels}
\title{Wrapper for function dealing with mortality models}
\usage{
doMortalityModels(data, x, y, data.type = c("qx", "mx", "dx"), ...)
}
\arguments{
\item{data}{Mortality data}

\item{x}{Vector of ages.}

\item{y}{Vector of years.}

\item{data.type}{Type of data in 'data' argument. Options: 
\code{"qx", "mx", "dx"}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Wrapper for function dealing with mortality models
}
\examples{
x = 0:110
y = 1960:1980
dxm <- dxForecast::dxForecast.data$dx$male[paste(x), paste(y)]
dx  <- apply(dxm, 2, function(x) x/sum(x))

M <- doMortalityModels(data = dx, x, y, data.type = "dx")
P <- doForecasts(M, h = 16, ci = 95, jumpchoice = "actual")

oex <- getObserved(M, type = "ex")
fex <- getFitted(M, type = "ex")
rex <- getResiduals(M, type = "ex")
pex <- getForecasts(P, type = "ex")
}
