% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMortalityModels.R
\name{doMortalityModels}
\alias{doMortalityModels}
\title{Fit Multiple Stochastic Mortality Models}
\usage{
doMortalityModels(data, x = NULL, y = NULL, data.in = c("qx", "mx",
  "dx", "lx"), models = c("MRWD", "LC"), verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{data.in}{Specify the type of input \code{data}. Various life table 
indices are accepted: \code{"qx", "mx", "dx", "lx"}.}

\item{models}{Mortality models to be evaluated.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Fit Multiple Stochastic Mortality Models
}
\examples{
x  <- 0:100
y  <- 2005:2016
D  <- MortalityForecast.data$dx[paste(x), paste(y)]
MM <- c("MRWD", "LeeCarter", "HyndmanUllah", "CoDa")

M <- doMortalityModels(data = D, x, y, data.in = "dx", models = MM)

oex <- getObserved(M, data.out = "ex")
fex <- getFitted(M, data.out = "ex")
rex <- getResiduals(M, data.out = "ex")
}
