% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMortalityModels.R
\name{doMortalityModels}
\alias{doMortalityModels}
\title{Wrapper for function dealing with mortality models}
\usage{
doMortalityModels(data, x, y, data.type = c("qx", "mx", "dx", "lx"),
  models = c("MRW", "MRWD", "LC", "FDM", "PLAT", "CoDa", "M4", "M4X",
  "M5", "M5X", "M6", "M6X"), exogen = NULL, ...)
}
\arguments{
\item{data}{Mortality data}

\item{x}{Vector of ages.}

\item{y}{Vector of years.}

\item{data.type}{Type of data in 'data' argument. Options: 
\code{"qx", "mx", "dx", "lx"}.}

\item{models}{Mortality models to be evaluated.}

\item{exogen}{Inlusion of exogenous variables.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Wrapper for function dealing with mortality models
}
\examples{
x = 0:100
y = 2005:2016
D  <- MortalityForecast.data$dx[paste(x), paste(y)]
ex <- MortalityForecast.data$ex
exogen <- ex[paste(y)]

MM = c("LC", "FDM", "CoDa", "M6", "M6X")
M <- doMortalityModels(data = D, x, y, data.type = "dx",
                       models = MM, exogen = exogen)

oex <- getObserved(M, what = "ex")
fex <- getFitted(M, what = "ex")
rex <- getResiduals(M, what = "ex")
}
