% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMortalityModels.R
\name{doMortalityModels}
\alias{doMortalityModels}
\title{Wrapper for function dealing with mortality models}
\usage{
doMortalityModels(data, x, y, data.type = c("qx", "mx", "dx", "lx"),
  exogen = NULL, ...)
}
\arguments{
\item{data}{Mortality data}

\item{x}{Vector of ages.}

\item{y}{Vector of years.}

\item{data.type}{Type of data in 'data' argument. Options: 
\code{"qx", "mx", "dx", "lx"}.}

\item{exogen}{Inlusion of exogenous variables.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Wrapper for function dealing with mortality models
}
\examples{
x = 0:100
y = 2005:2016
dxm <- dxForecast::dxForecast.data$dx$male[paste(x), paste(y)]
ex <- dxForecast::dxForecast.data$ex$male
exogen <- ex[paste(y)]

M <- doMortalityModels(data = dxm, x, y, data.type = "dx", exogen = exogen)

oex <- getObserved(M, what = "ex")
fex <- getFitted(M, what = "ex")
rex <- getResiduals(M, what = "ex")
}
