% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRanking.R
\name{doRanking}
\alias{doRanking}
\title{Rank models based on the accuracy or robustness results}
\usage{
doRanking(A)
}
\arguments{
\item{A}{Table containing accuracy or robustness measures.}
}
\description{
Rank models based on the accuracy or robustness results
}
\examples{
# For examples go to ?doBackTesting or ?doBBackTesting
}
\seealso{
\code{\link{doBackTesting}} 
\code{\link{doBBackTesting}}
}
\author{
Marius D. Pascariu
}
