% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalAccuracy.R
\name{evalAccuracy}
\alias{evalAccuracy}
\title{Generic for Computing Accuracy Measure from the \code{BackTesting} and 
\code{BBackTesting} objects;}
\usage{
evalAccuracy(object, ...)
}
\arguments{
\item{object}{An object of the class \code{BackTesting} or 
\code{BBackTesting};}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Generic for Computing Accuracy Measure from the \code{BackTesting} and 
\code{BBackTesting} objects;
}
\examples{
# For examples go to ?do.BackTesting and ?do.BBackTesting
}
\author{
Marius D. Pascariu
}
\keyword{internal}
