% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_fit.R
\name{find_ages_and_years}
\alias{find_ages_and_years}
\title{Find age and year vectors}
\usage{
find_ages_and_years(data, x, y)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}
}
\description{
Find age and year vectors
}
\keyword{internal}
