% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-models.R
\name{fitHyndmanUllah}
\alias{fitHyndmanUllah}
\title{Functional Demographic Model}
\usage{
fitHyndmanUllah(data, x, y, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Functional Demographic Model
}
\seealso{
\code{\link{fitLeeCarter}}
\code{\link{coda}}
}
