% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{get_dx_values}
\alias{get_dx_values}
\title{#' Get d[x] values based on k[t] forecast}
\usage{
get_dx_values(object, kt, jumpchoice, y, adj = 1)
}
\arguments{
\item{object}{An object of class \code{LeeCarter}.}

\item{kt}{Estimated kt vector of parameters in the model;}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{adj}{Adjustment to be used in the Li-Lee model only.}
}
\description{
#' Get d[x] values based on k[t] forecast
}
\keyword{internal}
