% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter2.R
\name{get_mx_values}
\alias{get_mx_values}
\title{Get m[x] values based on k[t] forecast
In this function we compute the m[x] values based on the extrapolation of
the k[t] time-series. If necesary an adjustment for the jump-off is 
provided.}
\usage{
get_mx_values(object, kt, jumpchoice, y, LL_adjustment = 0)
}
\arguments{
\item{object}{An object of class \code{LeeCarter2}.}

\item{kt}{Estimated kt vector of parameters in the Lee-Carter model;}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{LL_adjustment}{Adjustment to be used in the Li-Lee model.}
}
\description{
Get m[x] values based on k[t] forecast
In this function we compute the m[x] values based on the extrapolation of
the k[t] time-series. If necesary an adjustment for the jump-off is 
provided.
}
\keyword{internal}
