% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitted.R
\name{getFitted}
\alias{getFitted}
\title{Get Fitted Values}
\usage{
getFitted(object, data.out = c("qx", "mx", "dx", "lx", "Lx", "Tx", "ex"),
  ...)
}
\arguments{
\item{object}{An object of class \code{MortalityModels}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Get Fitted Values
}
