% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fitMortality.R
\name{getForecasts}
\alias{getForecasts}
\title{Get Predicted Values}
\usage{
getForecasts(object, type = c("qx", "mx", "dx", "lx", "Lx", "Tx", "ex"), ...)
}
\arguments{
\item{object}{An object of class \code{predict.MortalityModels}}

\item{type}{Specify the type of values to be extracted.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Get Predicted Values
}
