% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForecasts.R
\name{getForecasts}
\alias{getForecasts}
\title{Get Predicted Values}
\usage{
getForecasts(object, data.out = c("qx", "mx", "dx", "lx", "Lx", "Tx",
  "ex"), ...)
}
\arguments{
\item{object}{An object of class \code{doForecasts}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Get Predicted Values
}
\seealso{
\code{\link{doForecasts}}
}
\author{
Marius D. Pascariu
}
