% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMortalityModels.R
\name{getObserved}
\alias{getObserved}
\title{Get Observed Values}
\usage{
getObserved(object, what = c("qx", "mx", "dx", "lx", "Lx", "Tx", "ex"),
  ...)
}
\arguments{
\item{object}{An object of the class \code{MortalityModels}.}

\item{what}{Specify the type of values to be extracted.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Get Observed Values
}
