% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResiduals.R
\name{getResiduals}
\alias{getResiduals}
\title{Get Deviance Residuals}
\usage{
getResiduals(object, data.out = c("qx", "mx", "dx", "lx", "Lx", "Tx",
  "ex"), ...)
}
\arguments{
\item{object}{An object of class \code{MortalityModels}.}

\item{data.out}{Specify the type of data to be returned in output. 
Various life table indices are accepted: 
\code{"qx", "mx", "dx", "lx", "Lx", "Tx", "ex"}.}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Get Deviance Residuals
}
