% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_Plot_fit.R
\name{ggplotDistribConvergence}
\alias{ggplotDistribConvergence}
\title{Plot Convergence in Age at Death Distribution}
\usage{
ggplotDistribConvergence(mat, x, ny, level)
}
\arguments{
\item{mat}{Matrix containing the observed or fitted value.}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}
}
\description{
Plot Convergence in Age at Death Distribution
}
\keyword{internal}
