% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LiLee.R
\name{model_LiLee}
\alias{model_LiLee}
\title{The Li-Lee Mortality Model}
\usage{
model_LiLee(data, data.B, x = NULL, y = NULL, verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame or a matrix containing mortality data 
with ages \code{x} as row and time \code{y} as column.}

\item{data.B}{A data.frame or a matrix containing mortality data for the 
benchmark population. Must be the same format as in \code{data};}

\item{x}{Numerical vector indicating the ages in input \code{data}. Optional.
Default: \code{NULL}.}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
The output is a list with the components:
 \item{input}{List with arguments provided in input. Saved for convenience;}
 \item{info}{Short details about the model;}
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given 
 arguments;}
 \item{coefficients}{Estimated coefficients;}
 \item{fitted.values}{Fitted values of the estimated model;}
 \item{observed.values}{The observed values used in fitting arranged in the 
 same format as the fitted.values;}
 \item{residuals}{Deviance residuals;} 
 \item{x}{Vector of ages used in the fitting;} 
 \item{y}{Vector of years used in the fitting;} 
 \item{benchmark}{An object of class \code{LeeCarter} containing the fitted
 model for the benchmark population.}
}
\description{
Fit the Li-Lee mortality model
}
\details{
\insertNoCite{li2005}{MortalityForecast}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{predict.LiLee}}
}
