% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MEM.R, R/plot_Oeppen.R
\name{plot.MEM}
\alias{plot.MEM}
\alias{plot.Oeppen}
\alias{plot.OeppenC}
\title{Plot the Observed and Fitted Age-at-Death Distribution}
\usage{
\method{plot}{MEM}(x, plotType = c("fitted", "observed"), ny = 7,
  level = 80, ...)

\method{plot}{Oeppen}(x, plotType = c("fitted", "observed"), ny = 7,
  level = 80, ...)

\method{plot}{OeppenC}(x, plotType = c("fitted", "observed"), ny = 7,
  level = 80, ...)
}
\arguments{
\item{x}{A fitted mortality model.}

\item{plotType}{The type of the plot. The alternatives are 
\code{"fitted"}, \code{"observed"}. Default: \code{"fitted"}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}

\item{level}{Sets the number of quantiles the data should be broken into. 
Default: 80.}

\item{...}{Other plotting parameters to be passed to the plotting 
functions. This can be used to control the appearance of the plots.}
}
\description{
Plot the observed and fitted age-at-death distribution returned by a 
\code{Oeppen, OeppenC} or \code{MEM} model.
}
\details{
Note: the output is a ggplot2 object. Therefore, one can add or 
modify the components of the figure.
}
\examples{
# For examples go to ?model.Oeppen or ?model.OeppenC or ?model.MEM
}
\seealso{
\code{\link{model.Oeppen}}
\code{\link{model.OeppenC}}
\code{\link{model.MEM}}
}
