% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_Plot_predict.R
\name{plot.predict.MEM}
\alias{plot.predict.MEM}
\title{ggplot the predicted values of a Maximum-Entropy Mortality Model}
\usage{
\method{plot}{predict.MEM}(x, y = NULL, plotType = c("mean", "lower",
  "upper", "raw_moments", "normalised_moments"), ny = 7, quant = c(0.1,
  0.9), ...)
}
\arguments{
\item{x}{An object of the class \code{\link{predict.MEM}}.}

\item{y}{An object of the class \code{\link{MEM}}. 
Needed only for ploting moments. See \code{plotType}.}

\item{plotType}{The type of the plot. The alternatives are 
\code{"mean", "lower", "upper", "raw_moments", "normalised_moments"}. 
Default: \code{"mean"}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}

\item{quant}{Sets the number of quantiles the data should be broken into. 
Default: c(0.1, 0.9).}

\item{...}{Other plotting parameters to be passed to the plotting 
functions. This can be used to control the appearance of the plots.}
}
\description{
ggplot the predicted values of a Maximum-Entropy Mortality Model
}
\examples{
x  <- 0:110
y  <- 1965:2014
dx <- MortalityForecast.data$dx[paste(x), paste(y)]
M  <- MEM(dx, x, y, n = 5)  
P  <- predict(M, h = 16, x.h = 0:120)

plot(P, plotType = "mean")
plot(P, plotType = "lower")
plot(P, plotType = "upper")

plot(P, M, plotType = "raw_moments")
plot(P, M, plotType = "normalised_moments")
}
