% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MEM_plot.R
\name{plot.predict.MEM}
\alias{plot.predict.MEM}
\title{ggplot the predicted values of a Maximum-Entropy Mortality Model}
\usage{
\method{plot}{predict.MEM}(x, y = NULL, plotType = c("mean", "lower",
  "upper", "raw_moments", "normalised_moments", "scaled_moments"),
  ny = 7, level = 80, ...)
}
\arguments{
\item{x}{An object of the class \code{\link{predict.MEM}}.}

\item{y}{An object of the class \code{\link{model_MEM}}. 
Needed only for ploting moments. See \code{plotType}.}

\item{plotType}{The type of the plot. The alternatives are 
\code{"mean", "lower", "upper", "raw_moments", "normalised_moments", "scaled_moments"}. 
Default: \code{"mean"}.}

\item{ny}{Number of years to be selected from input data and to be added in the plot.}

\item{level}{Sets the number of quantiles the data should be broken into. 
Default: 80.}

\item{...}{Other plotting parameters to be passed to the plotting 
functions. This can be used to control the appearance of the plots.}
}
\description{
ggplot the predicted values of a Maximum-Entropy Mortality Model
}
\examples{
# For examples go to ?predict.MEM
}
