% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_HyndmanUllah.R
\name{predict.HyndmanUllah}
\alias{predict.HyndmanUllah}
\title{Forecast age-specific death rates using the Hyndman-Ullah mortality model}
\usage{
\method{predict}{HyndmanUllah}(object, h, level = 95,
  jumpchoice = c("actual", "fit"), verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{HyndmanUllah}.}

\item{h}{Number of years to be forecast in the future.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Additional arguments to be passed to \code{\link[forecast]{Arima}}}
}
\value{
The output is an object of class \code{"predict.Oeppen"} with the components:
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given arguments;}
 \item{predicted.values}{A list containing the predicted values together
 with the associated prediction intervals given by the estimated 
 model over the forecast horizon \code{h};}
 \item{kt.arima}{An object of class \code{ARIMA} that contains all the
 components of the fitted time series model used in \code{kt} prediction;} 
 \item{kt}{The extrapolated \code{kt} parameters;}
 \item{conf.intervals}{Confidence intervals for the extrapolated \code{kt} 
 parameters;}
 \item{x}{Vector of ages used in prediction;} 
 \item{y}{Vector of years used in prediction;}
 \item{info}{Short details about the model.}
}
\description{
Forecast age-specific death rates using the Hyndman-Ullah mortality model
}
\examples{
# Data
x  <- 0:89
y  <- 1985:2014
mx <- HMD_male$mx$GBRTENW[paste(x), paste(y)]

M <- fit_HyndmanUllah(data = mx, x = x, y = y) # fit
P <- predict(M, h = 20)  # forecast
P
}
\seealso{
\code{\link{fit_HyndmanUllah}}
}
\author{
Marius D. Pascariu and Marie-Pier Bergeron-Boucher
}
