% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LiLee.R
\name{predict.LiLee}
\alias{predict.LiLee}
\title{Forecast age-specific death rates using the Li-Lee model}
\usage{
\method{predict}{LiLee}(object, h, order = c(0, 1, 0),
  include.drift = TRUE, level = c(80, 95), jumpchoice = c("actual",
  "fit"), method = "ML", verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{LiLee}.}

\item{h}{Number of years to be forecast in the future.}

\item{order}{A specification of the non-seasonal part of the ARIMA model: 
the three components (p, d, q) are the AR order, the degree of differencing, 
and the MA order. If \code{order = NULL}, the ARIMA order will be estimated 
automatically using the KPPS algorithm.}

\item{include.drift}{Logical. Should the ARIMA model include a linear drift 
term? If \code{include.drift = NULL}, the model will be estimated 
automatically.}

\item{level}{Significance level of the confidence interval.}

\item{jumpchoice}{Method used for computation of jumpchoice. 
Possibilities: \code{"actual"} (use actual rates from final year) 
and \code{"fit"} (use fitted rates).}

\item{method}{ARIMA fitting method: maximum likelihood or minimize 
conditional sum-of-squares. Options to use: conditional-sum-of-squares 
(\code{"CSS-ML"}), maximum likelihood (\code{"ML"}) and \code{"CSS"}.}

\item{verbose}{A logical value. Set \code{verbose = FALSE} to silent 
the process that take place inside the function and avoid progress messages.}

\item{...}{Additional arguments to be passed to \code{\link[forecast]{Arima}}}
}
\value{
The output is a list with the components:
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given 
 arguments;}
 \item{predicted.values}{A list containing the predicted values together
 with the associated prediction intervals given by the estimated 
 model over the forecast horizon \code{h};}
 \item{conf.intervals}{Confidence intervals for the extrapolated \code{kt} 
 parameters;}
 \item{kt.arima}{An object of class \code{ARIMA} that contains all the
 components of the fitted time series model used in \code{kt} prediction;} 
 \item{kt}{The extrapolated \code{kt} parameters;}
 \item{x}{Vector of ages used in prediction;} 
 \item{y}{Vector of years used in prediction;}
 \item{info}{Short details about the model;}
 \item{benchmark}{An object containing the predicted results for the 
 benchmark population.}
}
\description{
Forecast age-specific death rates using the Li-Lee model
}
\examples{
# Data
x <- 0:89
y <- 1985:2014
B.mx <- HMD_male$mx$USA[paste(x), paste(y)]
mx <- HMD_male$mx$GBRTENW[paste(x), paste(y)]

M <- model_LiLee(data = mx, B.data = B.mx, x = x, y = y) # fit
P <- predict(M, h = 20)  # forecast
P
}
\seealso{
\code{\link{model_LiLee}}
}
\author{
Marius D. Pascariu and Marie-Pier Bergeron-Boucher
}
