% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MRW.R
\name{predict.MRW}
\alias{predict.MRW}
\title{Forecast age-specific death rates using the a Multivariate Random Walk Model}
\source{
The original implementation of this function was taken from 
\code{StMoMo} R package.
}
\usage{
\method{predict}{MRW}(object, h = 10, level = c(80, 95), ...)
}
\arguments{
\item{object}{An object of class \code{MRW}.}

\item{h}{Number of years to be forecast in the future.}

\item{level}{Significance level of the confidence interval.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
An object of the class \code{predict.MRW} with components:
 \item{call}{An unevaluated function call, that is, an unevaluated 
 expression which consists of the named function applied to the given 
 arguments;}
 \item{info}{Short details about the model;}
 \item{predicted.values}{data.frame with the central forecast;}
 \item{conf.intervals}{List with lower and upper limits for prediction 
 intervals;}
 \item{x}{Numerical vector indicating the age classes in output;}
 \item{y}{Numerical vector indicating the years in output.}
}
\description{
Forecast age-specific death rates using the a Multivariate Random Walk Model
}
\details{
\insertNoCite{haberman2011}{MortalityForecast}
}
\examples{
# For examples go to ?model.MRW
}
\references{
\insertAllCited{}
}
\author{
Marius D. Pascariu
}
