% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRW.R
\name{predict.MRW}
\alias{predict.MRW}
\title{Forecast a Multivariate Random Walk Model}
\usage{
\method{predict}{MRW}(object, h = 10, level = c(80, 95), ...)
}
\arguments{
\item{object}{An object of the class \code{"MWR"}.}

\item{h}{Number of periods for forecasting.}

\item{level}{Confidence level for prediction intervals.}

\item{...}{Ignored.}
}
\value{
An object of the class \code{"predict.MRW"} with components:
\item{mean}{ Array with the central forecast.}
\item{lower}{ Three dimensional array with lower limits for prediction intervals.}
\item{upper}{ Three dimensional array with upper limits for prediction intervals.}
}
\description{
Returns forecasts and other information for a Multivariate 
Random Walk Model.
}
