% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MEM.R
\name{print.MEM}
\alias{print.MEM}
\title{Print function for Maximum-Entropy Mortality Model}
\usage{
\method{print}{MEM}(x, ...)
}
\arguments{
\item{x}{An object of class \code{MEM}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for Maximum-Entropy Mortality Model
}
\keyword{internal}
