% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMortalityModels.R
\name{print.MortalityModels}
\alias{print.MortalityModels}
\title{Print function for doMortalityModels}
\usage{
\method{print}{MortalityModels}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"MortalityModels"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for doMortalityModels
}
\keyword{internal}
