% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R
\name{print.predict.LeeCarter2}
\alias{print.predict.LeeCarter2}
\title{Print predict.Oeppen}
\usage{
\method{print}{predict.LeeCarter2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.LeeCarter2"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print predict.Oeppen
}
\keyword{internal}
