% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_default}
\alias{print_default}
\title{Print function for mortality models}
\usage{
print_default(x, ...)
}
\arguments{
\item{x}{A fitted mortality model}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for mortality models
}
\keyword{internal}
