% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace.zeros}
\alias{replace.zeros}
\title{If there are death rates are equal to zero they will be replaced
using the multiplicative replacement strategy.}
\usage{
replace.zeros(mx, radix = 1e+05)
}
\arguments{
\item{mx}{Matrix of death rates}

\item{radix}{radix. Default: 1e5.}
}
\description{
If there are death rates are equal to zero they will be replaced
using the multiplicative replacement strategy.
}
\examples{
x  <- 0:25
y  <- 2005:2016
mx <- HMD_male$mx$DNK[paste(x), paste(y)]
mx == 0

new.mx <- replace.zeros(mx)
new.mx == 0

sum(mx) == sum(new.mx)
(mx - new.mx)/mx * 100
}
