% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_HyndmanUllah.R
\name{residuals.HyndmanUllah}
\alias{residuals.HyndmanUllah}
\title{Residuals of the Hyndman-Ullah Mortality Model}
\usage{
\method{residuals}{HyndmanUllah}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"HyndmanUllah"}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Residuals of the Hyndman-Ullah Mortality Model
}
\examples{
# For examples go to ?model.HyndmanUllah
}
