% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MEM.R
\name{residuals.MEM}
\alias{residuals.MEM}
\title{Residuals of the Maximum-Entropy Mortality Model}
\usage{
\method{residuals}{MEM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MEM};}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Computed deviance residuals for a fitted Maximum-Entropy Mortality Model.
}
\examples{
# For examples go to ?model.MEM
}
