% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Oeppen.R
\name{residuals.Oeppen}
\alias{residuals.Oeppen}
\title{Residuals of the CoDa-LC Mortality Model}
\usage{
\method{residuals}{Oeppen}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"Oeppen"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Residuals of the CoDa-LC Mortality Model
}
