% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals_default}
\alias{residuals_default}
\title{Extract Model Residuals}
\usage{
residuals_default(object, ...)
}
\arguments{
\item{object}{A fitted mortality model}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Extract Model Residuals
}
\keyword{internal}
