% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Residuals.R
\name{summary.get.Residuals}
\alias{summary.get.Residuals}
\title{Summary for get.Residuals}
\usage{
\method{summary}{get.Residuals}(object, ..., digits = max(4L,
  getOption("digits") - 2L))
}
\arguments{
\item{object}{An object of class \code{\link{get.Residuals}}.}

\item{...}{Arguments to be passed to or from other methods.}

\item{digits}{Number of digits to display.}
}
\description{
Summary for get.Residuals
}
\keyword{internal}
