% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResiduals.R
\name{summary.getResiduals}
\alias{summary.getResiduals}
\title{Summary for getResiduals}
\usage{
\method{summary}{getResiduals}(object, ..., digits = max(3L,
  getOption("digits") - 3L))
}
\arguments{
\item{object}{An object of the class \code{\link{getResiduals}}.}

\item{...}{Arguments to be passed to or from other methods.}

\item{digits}{number of digits to display.}
}
\description{
Summary for getResiduals
}
