% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LeeCarter.R, R/model_LiLee.R,
%   R/model_Oeppen.R, R/model_OeppenC.R
\name{summary.LeeCarter}
\alias{summary.LeeCarter}
\alias{summary.LiLee}
\alias{summary.Oeppen}
\alias{summary.OeppenC}
\title{Generic Summary}
\usage{
\method{summary}{LeeCarter}(object, ...)

\method{summary}{LiLee}(object, ...)

\method{summary}{Oeppen}(object, ...)

\method{summary}{OeppenC}(object, ...)
}
\arguments{
\item{object}{A fitted mortality model}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Generic Summary
}
\keyword{internal}
