% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wide2long}
\alias{wide2long}
\title{Wide table to long table}
\usage{
wide2long(data, x = NULL, y = NULL, which.x = NULL, which.y = NULL,
  ...)
}
\arguments{
\item{data}{Wide table}

\item{x}{vector of ages}

\item{y}{Numerical vector indicating the years in input \code{data}. Optional.
Default: \code{NULL}.}

\item{which.x}{subset x vector}

\item{which.y}{subset y vector}

\item{...}{Arguments to be passed to or from other methods.}
}
\description{
Wide table to long table
}
\keyword{internal}
